Cross Validation for Ensemble Sparse Partial Least Squares Regression
Source: R/cv.enspls.R

K-fold cross validation for ensemble sparse partial least squares regression.

cv.enspls(x, y, nfolds = 5L, verbose = TRUE, ...)

Arguments

x
Predictor matrix.

y
Response vector.

nfolds
Number of cross-validation folds, default is 5. Note that this is the CV folds for the ensemble sparse PLS model, not the individual sparse PLS models. To control the CV folds for single sparse PLS models, please use the argument cvfolds.

verbose
Shall we print out the progress of cross-validation?

...
Arguments to be passed to enspls.fit.

Value

A list containing:

ypred - a matrix containing two columns: real y and predicted y

residual - cross validation result (y.pred - y.real)

RMSE - RMSE

MAE - MAE

Rsquare - Rsquare

Note

To maximize the probablity that each observation can be selected in the test set (thus the prediction uncertainty can be measured), please try setting a large reptimes.

See also

Examples

# This example takes one minute to run

# NOT RUN {
data("logd1k")
x = logd1k$x
y = logd1k$y

set.seed(42)
cvfit = cv.enspls(x, y, reptimes = 10)
print(cvfit)
plot(cvfit)
# }