Ensemble Partial Least Squares Regression

Ensemble partial least squares regression.

enpls.fit(x, y, maxcomp = NULL, cvfolds = 5L, reptimes = 500L,
  method = c("mc", "boot"), ratio = 0.8, parallel = 1L)

Arguments

x

Predictor matrix.

y

Response vector.

maxcomp

Maximum number of components included within each model. If not specified, will use the maximum number possible (considering cross-validation and special cases where n is smaller than p).

cvfolds

Number of cross-validation folds used in each model for automatic parameter selection, default is 5.

reptimes

Number of models to build with Monte-Carlo resampling or bootstrapping.

method

Resampling method. "mc" (Monte-Carlo resampling) or "boot" (bootstrapping). Default is "mc".

ratio

Sampling ratio used when method = "mc".

parallel

Integer. Number of CPU cores to use. Default is 1 (not parallelized).

Value

A list containing all partial least squares model objects.

See also

See enpls.fs for measuring feature importance with ensemble partial least squares regressions. See enpls.od for outlier detection with ensemble partial least squares regressions.

Examples

data("alkanes")
x = alkanes$x
y = alkanes$y

set.seed(42)
fit = enpls.fit(x, y, reptimes = 50)
print(fit)

#> Coefficients of the Models by Ensemble Partial Least Squares
#> ---
#>                      [,1]         [,2]          [,3]          [,4]
#> (Intercept)    -0.1341451   -0.2551829  2.172304e-01    0.00857426
#> Chi.P.0      -624.3268693  127.9495281  1.758220e+05  102.79105792
#> Chi.P.1      -711.1068351  145.3089741 -1.579397e+05  115.46319431
#> Chi.P.2      2793.2357167   73.5525809 -2.805611e+04   83.45384635
#> Chi.P.3      1636.1470405  363.6447786 -3.349958e+03  309.52386541
#> Chi.P.4        -3.3283854   -2.4515710 -1.381826e+01   -7.09271174
#> Chi.P.5       327.0241700  184.7817126 -4.197605e+02   90.03433302
#> Chi.C.3        44.7380621   48.4043228 -3.965709e+01   16.84190848
#> Chi.PC.4     -166.0627908   48.2584083 -1.854417e+02  -32.44954351
#> Kappa.1      2763.4071402 -172.7346455 -4.053647e+04  -52.72226789
#> Kappa.2      -767.0086748  199.0131897 -1.641829e+05  135.54503485
#> Kappa.3     -2495.2861931   43.6960492 -1.976772e+04  105.83803849
#> Kappa.4      -409.5935423  133.9776812  2.146203e+05  109.54434145
#> Estate.1      -18.3846592 -142.8955057 -7.817916e+01 -116.54248730
#> Estate.2      -49.6094279   75.4050447  6.422195e+00   87.87171619
#> Estate.3        8.8360919    4.9752098  1.733758e+01   -2.55167502
#> MEDV.11      -408.3367607  -28.9076731 -5.427864e+02  -36.74223545
#> MEDV.12     -1386.1069078 -224.1186107 -1.713180e+03 -190.03969510
#> MEDV.13        31.7503595  -14.0054920  3.262546e+00   21.52682255
#> MEDV.22      -939.3290565 -438.8210811 -1.204974e+03 -225.30499948
#>                      [,5]         [,6]         [,7]          [,8]         [,9]
#> (Intercept)    0.07599117    0.1067077    0.1458698    0.09264158    0.1490362
#> Chi.P.0       85.53870530  140.2296856   63.9098564   84.80166078   73.7651435
#> Chi.P.1       86.57148368  151.7573150   64.6082960   84.43126114   73.9892941
#> Chi.P.2       87.09765320  143.8834531   62.6656392   96.12580664   75.1188822
#> Chi.P.3      110.67192304  370.1672256   73.3703103   83.47607124  100.2954347
#> Chi.P.4      -10.33671185   -4.8043218  -10.8745791  -10.29927258  -13.1536799
#> Chi.P.5       67.14164637   90.0563523  102.2164320   37.94902206   59.1031176
#> Chi.C.3       20.60374087   15.5231946   35.3315239   17.29538202   21.4295585
#> Chi.PC.4      13.89574546  -45.3765932   53.0209055   -8.28616557    8.8454150
#> Kappa.1        0.60812194 -112.8580062    2.5263587   24.70830501    7.8289700
#> Kappa.2       93.60476959  187.6450812   70.4254575   85.34256503   75.3403675
#> Kappa.3       85.72854186  136.3517466   58.7101104  102.54383097   91.8180882
#> Kappa.4       86.57153122  145.8406041   64.5425318   85.44570362   74.4054587
#> Estate.1    -144.79242514 -137.7549469   -5.1791429 -127.62197598 -118.0510873
#> Estate.2     116.08876003  122.5949775   21.5474014   91.39492702   91.0072613
#> Estate.3     -14.75301433    5.5260933  -17.6987535  -14.50760304  -14.1700272
#> MEDV.11      -31.81714452  -44.1167716  -25.8386879  -34.34194990  -36.0934408
#> MEDV.12     -121.00447148 -243.5754197 -101.8503170 -120.47046013 -116.8292342
#> MEDV.13       38.60399125   32.8296505    4.4031060   47.81787866   35.5617584
#> MEDV.22       38.46627561 -446.2187311   98.7572285   74.29956783   44.6388601
#>                    [,10]        [,11]        [,12]        [,13]        [,14]
#> (Intercept)    0.2279339    0.1153782    0.1220744   -0.1279214    0.1282585
#> Chi.P.0      122.7672770  133.2411155   62.4794447  123.7908006  129.3703597
#> Chi.P.1      132.8573535  144.9476858   62.7405511  145.3796504  137.3272971
#> Chi.P.2      106.8328120  117.5678820   62.7713150   53.9699047  135.3256614
#> Chi.P.3      310.2792267  457.7879541   70.7373020  470.1619789  366.0424976
#> Chi.P.4       -9.3914659   -8.6631138  -12.5217960   -4.9809252   -5.4268158
#> Chi.P.5      102.7095283  142.3987545   90.2642243  191.1530517  108.1835516
#> Chi.C.3       17.3657614   28.3584386   30.0864759   41.4213080   21.5603952
#> Chi.PC.4     -10.4252601  -19.5287236   49.1921231   16.9063538  -25.1072472
#> Kappa.1      -93.7751079 -150.6652027    6.3873630 -200.2769631  -90.8983087
#> Kappa.2      169.1699403  191.0794646   67.1299012  199.4119629  174.0434640
#> Kappa.3       94.7478910   89.4170991   65.3634940   64.9376687   77.0023359
#> Kappa.4      127.9856490  140.8509375   62.9554947  133.0194947  133.7728497
#> Estate.1    -111.2237881 -142.5154761    7.3755686 -161.3458191 -167.9020794
#> Estate.2      99.9519926  122.8729377   16.0442562  115.5783797  137.5003039
#> Estate.3       1.9231683   11.0884253  -15.2607303    9.9787404    5.2845534
#> MEDV.11      -35.1568269  -33.4734325  -23.2581580  -19.6077947  -41.8714281
#> MEDV.12     -218.7321145 -254.8669208 -106.8681650 -221.6285945 -239.4521708
#> MEDV.13       14.7532759    4.5807693    3.0572713  -14.1683422   27.5127112
#> MEDV.22     -343.1935347 -546.5437452  100.7089610 -516.4370192 -430.5229069
#>                     [,15]         [,16]         [,17]         [,18]
#> (Intercept)    0.05322363    0.04993734     0.2184389     0.1494825
#> Chi.P.0       62.11206445   83.05452886  -388.5995311  -553.8881628
#> Chi.P.1       62.60095269   84.21675875  -555.8344239  -902.4355687
#> Chi.P.2       61.26828082   82.39322595  2431.7343050  3101.6952977
#> Chi.P.3       74.56603415  114.67293755  1562.0132058  1694.4637232
#> Chi.P.4      -13.70395433  -10.22218255    -2.1626083    -6.5664666
#> Chi.P.5       85.74009750   64.21390424   248.4289473   298.4932541
#> Chi.C.3       28.98315740   19.26137994    19.6659440    33.1266988
#> Chi.PC.4      44.89457807   11.16846299  -236.1300782  -197.3967140
#> Kappa.1        3.65138534    3.37021534  3774.4527157  2218.4505430
#> Kappa.2       67.13331975   91.06298030  -724.7758548  -852.7937788
#> Kappa.3       61.17543396   78.47871272 -3024.4597515 -2128.8261457
#> Kappa.4       62.68731776   84.09430402  -269.4107678  -494.0541019
#> Estate.1     -10.41148291 -137.14002806    66.3295221   -19.8856874
#> Estate.2       8.87426087   96.10953435   -98.8427251   -12.2769899
#> Estate.3     -14.52208799  -17.25198412     6.7651541    11.2814109
#> MEDV.11      -22.11600431  -27.56241915  -545.6718355  -361.9981448
#> MEDV.12     -103.20602181 -109.70698363 -1750.1875097 -1233.7427985
#> MEDV.13        0.88403658   30.21143937    41.5906903    62.2354548
#> MEDV.22       99.88983020   61.87030551  -951.3962360  -946.6710917
#>                     [,19]         [,20]         [,21]        [,22]
#> (Intercept)    -0.1478895    0.08036854    0.04939338    0.2143056
#> Chi.P.0     -1720.3288686   78.73039324   64.48906361   87.7177338
#> Chi.P.1     -1509.0902948   80.02698048   65.22236097   88.3241635
#> Chi.P.2      5255.5572320   78.28193194   63.53411832   89.7479178
#> Chi.P.3      2371.5208499  113.47537252   73.27614641  117.7804790
#> Chi.P.4        -3.2523739  -10.12749416  -10.67345184  -10.4252273
#> Chi.P.5       417.4902434   58.32441129   97.26389965   55.6098518
#> Chi.C.3        44.6961726   15.15547941   34.23746338   18.7271573
#> Chi.PC.4     -242.2026971    5.65502484   53.45266034   -0.6252804
#> Kappa.1      2349.3892971    9.32930837    0.92359979    8.0173362
#> Kappa.2      -973.2764365   83.58344677   70.98965098   93.5255537
#> Kappa.3     -2469.8074279   87.29838465   56.27662424   92.9134999
#> Kappa.4     -1320.9014996   79.84561088   65.15342976   88.5871495
#> Estate.1       12.0660550  -98.51119906    0.48584203 -143.4825445
#> Estate.2      -64.0487883   60.79443277   12.17922418  110.4657291
#> Estate.3       10.8298369  -13.79172019  -18.26433995  -17.1064783
#> MEDV.11      -363.1322166  -29.31953771  -24.00847766  -29.2002516
#> MEDV.12     -1283.6036484 -116.29973841 -102.11083905 -116.9637526
#> MEDV.13       100.1357476   25.87757752    0.82220921   37.5136052
#> MEDV.22      -989.4261970   64.51826349   96.25160019   53.4656953
#>                     [,23]        [,24]       [,25]         [,26]         [,27]
#> (Intercept) -1.879464e-01    0.3296985   0.1006492 -5.740018e-02   -0.06867056
#> Chi.P.0     -4.682040e+02   60.0547082  58.8529279 -4.168074e+02  129.37667397
#> Chi.P.1     -8.329473e+02   60.2313887  59.1122071 -5.900644e+02  141.73137049
#> Chi.P.2      2.699312e+03   60.3322615  58.5842313  2.342335e+03  139.63726486
#> Chi.P.3      1.587679e+03   69.7470015  68.0076378  1.409523e+03  344.81392845
#> Chi.P.4     -5.509686e+00  -14.0809987 -11.2685777 -6.032723e+00   -4.09335495
#> Chi.P.5      3.167201e+02   83.0832830  87.0373257  2.916411e+02  123.06442421
#> Chi.C.3      4.042208e+01   30.0797156  30.1365962  4.056904e+01   25.58881693
#> Chi.PC.4    -1.525804e+02   48.4822264  47.1981815 -1.317254e+02   -6.88267267
#> Kappa.1      1.872925e+03    4.6581328   6.1482131  2.108194e+03  -84.93662296
#> Kappa.2     -5.100108e+02   63.7638483  62.8516004 -5.931769e+02  181.66250172
#> Kappa.3     -1.809228e+03   60.6075780  60.9622742 -1.932842e+03   65.63800363
#> Kappa.4     -5.704506e+02   60.4942475  59.2969053 -4.025159e+02  139.31995745
#> Estate.1    -7.534523e+01   -7.1156873 -10.2184271 -7.164823e+01 -159.43232256
#> Estate.2    -2.545101e-02   10.2965623   1.6158087  1.842837e+01  108.93986661
#> Estate.3     1.268478e+01  -13.7837040 -12.5867407  7.544991e+00    2.33844286
#> MEDV.11     -3.017087e+02  -23.9129530 -25.1270679 -3.152262e+02  -44.83831606
#> MEDV.12     -1.091796e+03 -102.7323191 -99.8746105 -1.123028e+03 -246.41018861
#> MEDV.13      3.905099e+01    0.7960565   4.3102997  3.599114e+01   20.56433698
#> MEDV.22     -9.274069e+02   95.3093917  99.3488251 -8.428254e+02 -402.81897840
#>                    [,28]         [,29]        [,30]         [,31]         [,32]
#> (Intercept)    0.1282971   -0.03913359   -0.1466288    -0.2497438  4.591004e-03
#> Chi.P.0      141.8362893  143.61509804  121.8141234  -349.2471405  8.401077e+01
#> Chi.P.1      149.9981235  157.02598051  132.8229954  -428.5845313  8.525340e+01
#> Chi.P.2      146.1180530  165.42480306   92.5286257  1994.8345464  8.285909e+01
#> Chi.P.3      358.8402245  256.23432526  393.4093443  1460.4255486  1.205857e+02
#> Chi.P.4       -4.6941276  -13.33592706   -4.8838271    -0.9365777 -9.625363e+00
#> Chi.P.5       98.4185567   73.28029741  150.2310342   319.5140601  6.266412e+01
#> Chi.C.3       14.6078177   12.08016196   30.9948422    47.2126412  2.016359e+01
#> Chi.PC.4     -33.9050794  -20.82332340   -0.2812580  -150.2896689  7.566164e+00
#> Kappa.1     -112.7479266  -66.39891649 -123.3992535  2951.1500248  2.794416e+00
#> Kappa.2      188.9898297  194.67781521  173.9054959  -578.0350124  9.139648e+01
#> Kappa.3      116.6565993  106.35086492   59.4807741 -2509.4252753  7.807440e+01
#> Kappa.4      147.2458294  152.09970320  128.8246784  -207.8835489  8.521083e+01
#> Estate.1    -135.5183499  -74.16433764 -128.7952149  -104.9171421 -1.768603e+02
#> Estate.2     114.5175209   57.94418009  106.3816480     6.0981330  1.345669e+02
#> Estate.3       6.3182363    4.19626240    7.8539361     6.5068318 -1.553030e+01
#> MEDV.11      -47.9623297  -57.81221969  -33.5084069  -416.8980997 -2.812434e+01
#> MEDV.12     -254.0333694 -270.17214685 -238.7348473 -1401.4051010 -1.112620e+02
#> MEDV.13       29.7957274   24.62887489    3.7645463    17.6294849  4.126864e+01
#> MEDV.22     -460.9506149 -432.04698423 -452.6683780  -906.1157586  4.648186e+01
#>                     [,33]       [,34]        [,35]        [,36]         [,37]
#> (Intercept)   -0.03217558  -0.1412318    0.1092737   -0.1879711 -2.609598e-02
#> Chi.P.0       87.55243084  61.0537545   62.0647079   11.0362431 -5.249033e+02
#> Chi.P.1       89.26189724  61.6897189   62.3086515  -27.7770747 -1.072812e+03
#> Chi.P.2       81.25142327  59.9099376   61.2902763  580.1546096  3.620401e+03
#> Chi.P.3      145.30350228  71.6385626   74.3484976  633.8468664  1.825875e+03
#> Chi.P.4      -10.85087499 -11.8690858  -12.9825032   -4.8093027 -7.612313e+00
#> Chi.P.5       54.92110170  95.1618129   88.2364175  233.8988729  3.189347e+02
#> Chi.C.3       13.28931307  28.1758025   28.8460046   49.9593216  3.491620e+01
#> Chi.PC.4      -6.83224063  51.4976443   48.6933991   15.1383708 -2.072744e+02
#> Kappa.1       -1.39807683   4.3084434    6.6297254  517.9580647  2.331681e+03
#> Kappa.2       96.61222505  66.9956362   65.9189655   -6.6552846 -9.650796e+02
#> Kappa.3       82.97668491  60.5792083   64.1341571 -546.0809533 -2.353554e+03
#> Kappa.4       88.61926081  61.6903446   62.5130428   14.8227032 -6.786657e+02
#> Estate.1    -155.99910092  -3.2354646   -6.5238045 -168.6994776  3.077736e+01
#> Estate.2     101.81753925   8.0797626    1.6473096   89.9549857 -6.252963e+01
#> Estate.3     -15.08006702 -18.1487762  -12.8844820    3.3833513  1.640641e+01
#> MEDV.11      -23.57393173 -26.1188899  -24.7339327 -107.6864951 -3.865513e+02
#> MEDV.12     -113.01876081 -99.7890219 -104.2899204 -463.6398909 -1.340629e+03
#> MEDV.13       32.55701619   2.8085820    1.4392485   -6.4563823  6.703480e+01
#> MEDV.22       35.61654537  96.9693122  103.4306512 -516.8886274 -1.073354e+03
#>                     [,38]         [,39]         [,40]         [,41]
#> (Intercept)  7.562105e-02   -0.09814751  3.978828e-02    0.01744246
#> Chi.P.0     -3.484450e+02  120.99907728 -4.495704e+02   63.14232214
#> Chi.P.1     -4.388642e+02  132.73856637 -6.931629e+02   64.00352745
#> Chi.P.2      1.916572e+03   97.67861896  2.790155e+03   61.23981996
#> Chi.P.3      1.165915e+03  379.53020228  1.605710e+03   77.96132776
#> Chi.P.4     -6.769486e+00   -4.29720097 -3.449886e+00  -10.51450926
#> Chi.P.5      2.518535e+02  130.14006721  3.026985e+02   82.76715239
#> Chi.C.3      3.716481e+01   28.03080976  3.889891e+01   25.67132028
#> Chi.PC.4    -9.852007e+01  -13.76229615 -1.656601e+02   47.07788833
#> Kappa.1      1.311660e+03 -100.63117826  2.752441e+03    0.84082486
#> Kappa.2     -3.980405e+02  169.03399776 -8.119793e+02   69.63357738
#> Kappa.3     -1.275078e+03   75.38925017 -2.444528e+03   56.18771976
#> Kappa.4     -2.741276e+02  127.99985230 -4.072771e+02   63.86733986
#> Estate.1    -1.110520e+02 -160.12725243 -6.626913e+01   -8.21869837
#> Estate.2     4.582515e+01  120.73957859 -6.638115e+00    3.76226373
#> Estate.3     8.249203e+00    2.56613960  6.070825e+00  -15.58573895
#> MEDV.11     -2.282686e+02  -36.61953628 -4.117971e+02  -22.70794902
#> MEDV.12     -8.329892e+02 -226.65142700 -1.391872e+03 -100.47446494
#> MEDV.13      3.529547e+01   17.09118819  5.304472e+01   -1.53128537
#> MEDV.22     -7.666075e+02 -377.79023774 -9.300474e+02  101.46154190
#>                    [,42]        [,43]         [,44]         [,45]         [,46]
#> (Intercept)    0.0531083    0.0966189    0.35359955 -5.924536e-02    0.02768534
#> Chi.P.0      112.7181867   93.2811480  145.26615984 -1.729435e+02  120.38434552
#> Chi.P.1      122.8187934   94.5116040  156.60925009 -1.198187e+02  128.53076356
#> Chi.P.2      125.4197214   91.4360326  121.66481881  1.156029e+03   84.30563205
#> Chi.P.3      371.8055199  140.2394021  412.42055562  1.060722e+03  414.27395939
#> Chi.P.4       -9.6137243   -5.7421998   -4.01177006 -6.636268e+00   -5.97467536
#> Chi.P.5      145.0570354   43.8073422  149.06380844  2.367928e+02  136.09793386
#> Chi.C.3       29.4700392   15.7957643   29.39233557  3.630365e+01   30.19204983
#> Chi.PC.4       5.4157812  -22.4650494   -5.25945751 -1.112636e+02  -29.79224983
#> Kappa.1      -86.4296122    8.2443069 -135.83332091  1.847393e+03 -102.15439750
#> Kappa.2      161.9827012   98.0110962  206.91648254 -3.062927e+02  159.04864803
#> Kappa.3       34.4768390  100.5479301   92.13164033 -1.535894e+03  106.36553884
#> Kappa.4      119.0147099   94.3468303  151.63108125 -8.889987e+01  124.08894957
#> Estate.1    -129.4728065 -208.3371127  -70.38340498 -1.066902e+02 -160.68055941
#> Estate.2      99.5311225  139.2530255   60.47011009  1.869657e+01  126.12080045
#> Estate.3       4.6720675  -17.1035517   10.94939452  8.871946e+00    4.01863304
#> MEDV.11      -31.2365027  -33.4831779  -53.14549329 -2.979312e+02  -45.04625485
#> MEDV.12     -231.4530219 -112.2252361 -291.09901818 -9.756917e+02 -236.35326322
#> MEDV.13        2.9738693   53.4433713    0.09676744  2.021013e+00   19.74048814
#> MEDV.22     -393.7899076   45.3831799 -563.57243731 -7.248701e+02 -393.94017577
#>                     [,47]         [,48]         [,49]         [,50]
#> (Intercept) -2.916929e-03  1.155435e-02   -0.01789776 -9.268915e-02
#> Chi.P.0      1.112991e+02 -1.750916e+02   51.16208948 -8.369769e+02
#> Chi.P.1      1.185525e+02 -5.713816e+02   82.51719300 -1.438619e+03
#> Chi.P.2      1.733605e+02  2.090983e+03  545.77463217  4.425125e+03
#> Chi.P.3      3.845523e+02  1.370427e+03  567.45498374  2.376641e+03
#> Chi.P.4     -6.808433e+00 -1.067089e+00   -2.28131216 -7.207404e+00
#> Chi.P.5      9.779089e+01  2.842443e+02  172.09538698  4.293297e+02
#> Chi.C.3      1.469998e+01  4.060617e+01   34.44132171  4.712337e+01
#> Chi.PC.4    -3.924689e+01 -1.524397e+02  -10.58040895 -2.568584e+02
#> Kappa.1     -5.484894e+01  2.213898e+03  336.76525437  3.280665e+03
#> Kappa.2      1.506492e+02 -5.744457e+02   86.21311057 -1.174485e+03
#> Kappa.3      6.367526e+01 -1.968968e+03 -363.26529339 -3.080339e+03
#> Kappa.4      1.219042e+02 -4.197794e+02   80.28188203 -1.002386e+03
#> Estate.1    -1.455921e+02 -1.070215e+02  -97.60897299 -5.485354e+01
#> Estate.2     1.287603e+02  1.978080e+01   56.12914530 -3.381215e+01
#> Estate.3     4.811437e+00  5.277743e+00   11.50856011  9.507037e+00
#> MEDV.11     -4.399306e+01 -3.405317e+02 -103.09630883 -4.715528e+02
#> MEDV.12     -2.496197e+02 -1.119025e+03 -466.67173235 -1.621943e+03
#> MEDV.13      2.879928e+01  3.619482e+01    6.18139835  7.226712e+01
#> MEDV.22     -4.076575e+02 -8.019454e+02 -664.34486620 -1.088681e+03
#>  [ reached getOption("max.print") -- omitted 2 rows ]

predict(fit, newx = x)

#>   [1] -2126.1900240 -2111.9578408 -1927.2518075 -1914.9849411 -1727.4005727
#>   [6] -1716.9209019 -1527.4813580 -1517.5585339 -1327.4601949 -1317.7143953
#>  [11] -1126.8346450 -1117.1826634 -1152.4079367  -926.3024411  -917.3150825
#>  [16]  -954.1921113  -726.0212947  -716.9764621  -755.3969700  -525.8672737
#>  [21]  -517.1304808  -556.6486407  -325.9746203  -317.1941670  -357.9228413
#>  [26]  -126.2044706  -117.6014683  -159.2414821    73.4786398    81.9942190
#>  [31]    68.4660664    59.0416609    52.4283183    48.9176501    40.2621153
#>  [36]    39.5798139   273.0787022   281.6042626   268.0145887   258.5688839
#>  [41]   251.9189755   239.4708384   238.4049795   472.4898595   481.0057313
#>  [46]   437.0405535  -682.8612931  -607.9960427  -484.8741774  -484.4769089
#>  [51]  -508.2361134  -505.2464335  -520.4118008  -385.7756880  -397.3441005
#>  [56]  -408.2534926  -416.2959667  -420.4863165  -281.5794303  -286.1197747
#>  [61]  -309.0049812  -308.0276454  -301.1545443  -320.2105078  -185.9245866
#>  [66]  -195.0877015  -198.6660049  -209.5480309  -220.0597254   -81.9814009
#>  [71]   -87.2246059  -110.4112864  -108.6779201  -120.0566490    13.8448271
#>  [76]     4.4614088     2.6827989    18.0973156     0.6985103   -17.7133785
#>  [81]   117.5690638   111.9389322   111.4428860    88.7032226    88.4855048
#>  [86]    77.0868064    73.1341621   207.5193738   201.1912364   181.6291214
#>  [91]   173.4054220   174.8724242   165.4479977   314.4542924   310.4397765
#>  [96]   287.1930239   287.5432076   277.0039380   272.3160552   403.4349923
#> [101]   396.4378296   380.8252369   372.5100062   369.4600940   364.1731744
#> [106]   516.3470752   509.4342034   488.3812498   486.0795625   475.7981749
#> [111]   471.5172541   600.7346131   585.6801799   567.0230418   708.5022326
#> [116]   687.7274104   684.6899515   666.7468645   784.4491057  -567.9017305
#> [121]  -479.5154097  -369.3398414  -253.1182939  -170.9846303   -55.0207931
#> [126]   -83.6194354     8.8055074   145.2178708   114.6181213   104.1762472
#> [131]    94.6740835   229.3013244   223.5516412   206.7088028   192.6293715
#> [136]   345.2123800   313.0206867   302.0840655   292.2716810   428.3897892
#> [141]   427.0429864   405.0187956   397.4141093   390.1606672   544.2693780
#> [146]   514.7046859   500.9640052   488.6703912   491.8581467   626.9763912
#> [151]   596.1135078   586.9230186   743.2616981   710.3462171   701.9023499
#> [156]   685.7877746   784.1930289   908.6180374   883.4590563   982.5243755
#> [161]   -28.3733401   168.9011079   152.2911179   368.7802071   351.5252652
#> [166]   567.7195755   524.9146588   518.2862374   518.0625258   631.7099745
#> [171]   620.3124658   615.5274678   766.5586671   723.0790492   716.1130040
#> [176]   714.8478938   813.1360767  -340.2225521  -150.9329026  -917.3073985
#> [181]  -683.8543337  -481.3423912  -508.6815778  -398.0556482  -415.1703458
#> [186]  -309.1427708  -321.2934956  -328.0735015  -197.7825808  -109.7393587
#> [191]    79.1530716    74.2976387   203.9927296   401.4461345   380.5822602
#> [196]  -251.6387991  -166.2844337  -100.5127479   116.5335456   103.3849931
#> [201]    97.2603717   320.3911382   302.5225468   295.5164059   406.3410413
#> [206]   405.8443277   392.4850064