Ensemble Sparse Partial Least Squares Regression

Ensemble sparse partial least squares regression.

enspls.fit(
  x,
  y,
  maxcomp = 5L,
  cvfolds = 5L,
  alpha = seq(0.2, 0.8, 0.2),
  reptimes = 500L,
  method = c("mc", "boot"),
  ratio = 0.8,
  parallel = 1L
)

Arguments

x
Predictor matrix.

y
Response vector.

maxcomp
Maximum number of components included within each model. If not specified, will use 5 by default.

cvfolds
Number of cross-validation folds used in each model for automatic parameter selection, default is 5.

alpha
Parameter (grid) controlling sparsity of the model. If not specified, default is seq(0.2, 0.8, 0.2).

reptimes
Number of models to build with Monte-Carlo resampling or bootstrapping.

method
Resampling method. "mc" (Monte-Carlo resampling) or "boot" (bootstrapping). Default is "mc".

ratio
Sampling ratio used when method = "mc".

parallel
Integer. Number of CPU cores to use. Default is 1 (not parallelized).

Value

A list containing all sparse partial least squares model objects.

See also

Author

Nan Xiao <https://nanx.me>

Examples

data("logd1k")
x <- logd1k$x
y <- logd1k$y

set.seed(42)
fit <- enspls.fit(
  x, y,
  reptimes = 5, maxcomp = 3,
  alpha = c(0.3, 0.6, 0.9)
)
print(fit)

#> Coefficients of the Models by Ensemble Sparse Partial Least Squares
#> ---
#>                             [,1]         [,2]         [,3]          [,4]
#> BalabanJ            -0.043615048 -0.070476160 -0.061234111 -0.0641194287
#> BertzCT             -0.010797407 -0.010380918 -0.014190111 -0.0198473677
#> Chi0                 0.034719388  0.031396910  0.031244743  0.0305194922
#> Chi0n                0.053961369  0.051389288  0.054495519  0.0524801055
#> Chi0v                0.053701577  0.051940514  0.053866466  0.0534589673
#> Chi1                 0.051776232  0.049156542  0.049348357  0.0492213616
#> Chi1n                0.070459184  0.069096237  0.072714905  0.0727060040
#> Chi1v                0.044723121  0.043485995  0.047373453  0.0501593866
#> Chi2n                0.044969369  0.049246878  0.053608909  0.0505554796
#> Chi2v               -0.007655968 -0.004618584 -0.001476135  0.0001540116
#> Chi3n                0.032584667  0.042587170  0.046107374  0.0415386369
#> Chi3v               -0.028277116 -0.022984653 -0.021257238 -0.0203092032
#> Chi4n                0.025069486  0.035056020  0.037250699  0.0377062265
#> Chi4v               -0.032725942 -0.028311655 -0.028345799 -0.0254204075
#> EState_VSA1         -0.145952322 -0.151990047 -0.160770607 -0.1525700022
#> EState_VSA10        -0.156900443 -0.168802906 -0.175239253 -0.1722320471
#> EState_VSA11        -0.127792171 -0.135963476 -0.143448807 -0.1529079518
#> EState_VSA2         -0.088764258 -0.081331379 -0.090132085 -0.0733057615
#> EState_VSA3          0.056790068  0.028421271  0.053648385  0.0386607294
#> EState_VSA4          0.063669587  0.074797750  0.074053361  0.0699041874
#> EState_VSA5          0.076840942  0.113331436  0.116703240  0.1013286787
#> EState_VSA6         -0.039520870 -0.054418432 -0.060441879 -0.0391363977
#> EState_VSA7          0.167588096  0.158416601  0.168272101  0.1618998057
#> EState_VSA8          0.009235939  0.004258386 -0.006701922 -0.0011150675
#> EState_VSA9          0.160639733  0.170068653  0.176254590  0.1714440988
#> ExactMolWt           0.050748220  0.048487236  0.047997890  0.0454753788
#> FractionCSP3        -0.002537413  0.016775756  0.014177552  0.0000000000
#> HallKierAlpha        0.019451372  0.000000000  0.000000000  0.0329693176
#> HeavyAtomCount       0.041665347  0.039779625  0.039215273  0.0389117918
#> HeavyAtomMolWt       0.049142955  0.047099309  0.045859934  0.0431088851
#> Ipc                  0.057240507  0.060688810  0.069024848  0.0694421064
#> Kappa1               0.038204987  0.032402483  0.032511070  0.0348008587
#> Kappa2               0.073454618  0.061351458  0.062538464  0.0699833246
#> Kappa3               0.068769152  0.053787212  0.048842217  0.0663792137
#> LabuteASA            0.050468473  0.048957140  0.047963605  0.0492392904
#> MaxAbsEStateIndex   -0.058201076 -0.046954270 -0.035764073 -0.0459235492
#> MaxEStateIndex      -0.058201076 -0.046954270 -0.035764073 -0.0459235492
#> MinAbsEStateIndex    0.050724276  0.044527421  0.044215313  0.0353840523
#> MinEStateIndex       0.080735399  0.074187812  0.072888633  0.0724654251
#> MolMR                0.064441013  0.060992387  0.061907090  0.0617893424
#> MolWt                0.050700489  0.048453503  0.047956295  0.0454522770
#> NumValenceElectrons  0.041901095  0.039028411  0.040413395  0.0398859275
#> PEOE_VSA1           -0.123885717 -0.111986458 -0.128388290 -0.1335502496
#> PEOE_VSA10           0.060135647  0.072095820  0.042111810  0.0611046246
#> PEOE_VSA11           0.027518533  0.000000000  0.054256529  0.0593889088
#> PEOE_VSA12          -0.070409371 -0.085305819 -0.078339953 -0.0775848326
#> PEOE_VSA13           0.071124088  0.073838570  0.087264498  0.0780882742
#> PEOE_VSA14          -0.059283402 -0.052884362 -0.059018175 -0.0639531460
#> PEOE_VSA2           -0.135631007 -0.136699485 -0.158100226 -0.1453559333
#> PEOE_VSA3           -0.015399543 -0.030536558 -0.023616048 -0.0258053476
#> PEOE_VSA4            0.089159585  0.082572159  0.101865179  0.1083556038
#> PEOE_VSA5            0.000000000  0.000000000  0.000000000  0.0000000000
#> PEOE_VSA6            0.252455157  0.257218113  0.264862957  0.3016795122
#> PEOE_VSA7            0.131757300  0.126559513  0.126970928  0.1081498903
#> PEOE_VSA8           -0.251591255 -0.244659553 -0.247195649 -0.2892456024
#> PEOE_VSA9            0.093581834  0.074355909  0.075437882  0.0803304497
#> SMR_VSA1            -0.045809417 -0.040299095 -0.050701107 -0.0532047159
#> SMR_VSA10            0.025671513  0.036747988  0.020329048  0.0236814814
#> SMR_VSA2             0.000000000  0.000000000  0.000000000  0.0000000000
#> SMR_VSA3            -0.196156845 -0.207343129 -0.205718468 -0.1978050051
#> SMR_VSA4             0.086514311  0.092000949  0.090979644  0.1135589734
#> SMR_VSA5             0.144746921  0.144103173  0.158124866  0.1643202372
#> SMR_VSA6            -0.086355302 -0.087439103 -0.084443620 -0.0980728660
#> SMR_VSA7             0.021803034  0.011365422 -0.006054515  0.0030932243
#> SMR_VSA9             0.164334759  0.153864976  0.180528755  0.1579959854
#> SlogP_VSA1          -0.045042683 -0.031673412 -0.087011109 -0.0688301411
#> SlogP_VSA10          0.119767503  0.115394919  0.103222595  0.0843745870
#> SlogP_VSA11          0.114675123  0.110717941  0.114633271  0.0846142758
#> SlogP_VSA12          0.103280911  0.095740332  0.101496317  0.1105220578
#> SlogP_VSA2          -0.207711269 -0.202806412 -0.200423733 -0.2013781453
#> SlogP_VSA3          -0.153164458 -0.151392486 -0.134059624 -0.1516867090
#> SlogP_VSA4           0.080989430  0.078481646  0.076140508  0.0736978562
#> SlogP_VSA5           0.174022630  0.173399624  0.191735830  0.2059561988
#> SlogP_VSA6           0.136169116  0.133023697  0.110695001  0.1313060684
#> SlogP_VSA7           0.077780951  0.068859412  0.083787537  0.0846865586
#> SlogP_VSA8           0.000000000  0.000000000  0.067225424  0.0563976493
#> TPSA                -0.220618737 -0.215277058 -0.233074352 -0.2170641606
#> VSA_EState10         0.093743521  0.084141143  0.095789368  0.1014189884
#> VSA_EState8          0.089001823  0.076620959  0.083263325  0.0777088146
#> VSA_EState9         -0.089786907 -0.081994686 -0.094262874 -0.0921319769
#>                             [,5]
#> BalabanJ            -0.073894883
#> BertzCT             -0.007687107
#> Chi0                 0.032319300
#> Chi0n                0.052919408
#> Chi0v                0.056854869
#> Chi1                 0.048305879
#> Chi1n                0.069904194
#> Chi1v                0.051215760
#> Chi2n                0.048557618
#> Chi2v                0.005050211
#> Chi3n                0.038590608
#> Chi3v               -0.016430846
#> Chi4n                0.028264216
#> Chi4v               -0.022830716
#> EState_VSA1         -0.151918402
#> EState_VSA10        -0.163913222
#> EState_VSA11        -0.114487283
#> EState_VSA2         -0.074990673
#> EState_VSA3          0.027238867
#> EState_VSA4          0.081673528
#> EState_VSA5          0.119173308
#> EState_VSA6         -0.062270377
#> EState_VSA7          0.169929249
#> EState_VSA8         -0.008508467
#> EState_VSA9          0.169614152
#> ExactMolWt           0.053132588
#> FractionCSP3         0.000000000
#> HallKierAlpha        0.000000000
#> HeavyAtomCount       0.039291679
#> HeavyAtomMolWt       0.051773716
#> Ipc                  0.040666065
#> Kappa1               0.035993825
#> Kappa2               0.066238470
#> Kappa3               0.056805777
#> LabuteASA            0.050833015
#> MaxAbsEStateIndex   -0.039681288
#> MaxEStateIndex      -0.039681288
#> MinAbsEStateIndex    0.058219064
#> MinEStateIndex       0.090836706
#> MolMR                0.065637066
#> MolWt                0.053101413
#> NumValenceElectrons  0.038876330
#> PEOE_VSA1           -0.131939533
#> PEOE_VSA10           0.063132746
#> PEOE_VSA11           0.000000000
#> PEOE_VSA12          -0.064913191
#> PEOE_VSA13           0.062907184
#> PEOE_VSA14          -0.053503851
#> PEOE_VSA2           -0.123394883
#> PEOE_VSA3           -0.021138001
#> PEOE_VSA4            0.044285049
#> PEOE_VSA5            0.000000000
#> PEOE_VSA6            0.269371381
#> PEOE_VSA7            0.118261626
#> PEOE_VSA8           -0.241824460
#> PEOE_VSA9            0.084102621
#> SMR_VSA1            -0.042033373
#> SMR_VSA10            0.054396319
#> SMR_VSA2             0.000000000
#> SMR_VSA3            -0.223644660
#> SMR_VSA4             0.086073286
#> SMR_VSA5             0.152844318
#> SMR_VSA6            -0.100483543
#> SMR_VSA7             0.017135164
#> SMR_VSA9             0.142266776
#> SlogP_VSA1          -0.055774803
#> SlogP_VSA10          0.082997064
#> SlogP_VSA11          0.088648633
#> SlogP_VSA12          0.112241959
#> SlogP_VSA2          -0.204690587
#> SlogP_VSA3          -0.156349706
#> SlogP_VSA4           0.077452104
#> SlogP_VSA5           0.198178112
#> SlogP_VSA6           0.127587128
#> SlogP_VSA7           0.000000000
#> SlogP_VSA8           0.000000000
#> TPSA                -0.214509241
#> VSA_EState10         0.084382079
#> VSA_EState8          0.077845374
#> VSA_EState9         -0.085217775

predict(fit, newx = x)

#>    [1] -0.21812478  0.12954374 -0.22836445  0.72266012 -0.65113402 -0.67093505
#>    [7]  0.61382930 -0.45282799  0.04079630 -0.06416978  0.79365384  0.12145376
#>   [13]  0.94735040 -0.31389461  0.71790961  0.86124578  0.11327756 -0.75816147
#>   [19] -0.14327956  0.50469871  1.13848250  0.13954518  1.82076630  1.01476461
#>   [25]  0.61597016  0.56588233  1.24949794  0.78184040  0.07023968 -0.32560716
#>   [31] -0.24545686  0.91085846  0.13738773 -0.53752609  0.63554844  0.51586274
#>   [37]  0.85954447 -0.34395002 -0.34058240  0.78555752 -0.37085733  0.90052950
#>   [43]  0.61609279 -0.07109915 -0.02722697  0.10367080  0.33241944  0.22370098
#>   [49] -0.18123400 -0.06948561 -0.04929142 -0.02625836  1.41948475  1.05831350
#>   [55]  0.06012419  0.29274788  0.08855050  1.19392450 -0.03963956 -0.03859852
#>   [61]  1.55698261 -0.48803906  1.02916294  1.73890966  1.87076311  1.37864665
#>   [67]  0.35687448  1.29920486  0.83581663  0.56929089  0.03734973 -0.08841345
#>   [73]  0.64824435  0.48630316  0.60329239  0.88549938  0.78232696  2.06775912
#>   [79]  0.16023107  0.21624428  0.65925599  0.28744740  0.35358473  0.02157764
#>   [85]  0.01210593  1.43460139  0.34611658  1.93062414  1.47011892  0.32332532
#>   [91] -2.75216196 -0.09239403  1.93228703  1.18745000  0.92896916  0.07936765
#>   [97]  0.57574740  3.06349960  1.57459052  0.69355876  2.03717644  2.32170182
#>  [103]  2.37872621  0.55675433  0.73749034  0.30514075  2.52839101  0.81552872
#>  [109]  1.86824783  2.05754038  1.89541525  0.95149315  2.16796051  1.50048941
#>  [115] -0.74991974  1.47102994  0.78846108  0.66757694  2.02478816  2.05613690
#>  [121]  2.15884081  0.96573182  1.45225310  2.03433230  0.69620378  0.67020676
#>  [127]  0.60352076  0.56561324  0.67297570  0.92097974  1.72925411  0.41515221
#>  [133]  0.69226657  1.00435210  3.18574778  1.65347736  2.00896410  0.88105080
#>  [139]  1.87852077  1.50276630  2.01475491  1.50202430  2.34787066  2.29547931
#>  [145]  1.73951951  1.87369573  0.10181703  1.60406984  2.63131115  1.67713625
#>  [151]  2.37758761  1.16420952  0.09014479  0.79038294  0.60273235  2.97783189
#>  [157]  1.55249758  1.27479322 -0.08177682  2.06224907  1.71203469  0.40430446
#>  [163]  1.03784594 -0.05585030  2.90963566  2.25324795  3.27452491  1.81465273
#>  [169]  0.06877732  3.14362520  0.96626930  0.09014479  2.25611724  1.53916617
#>  [175]  0.96493001  0.55419333  0.61360396  3.08944475  0.87403643  0.72478628
#>  [181]  1.04688102  1.16477997  2.57168201  0.42632452  2.01035298  0.76253155
#>  [187]  3.09110582  1.90929431  2.15958110  0.78591758  2.46880186  2.84296075
#>  [193]  1.28366233  2.85257488  2.68044169  3.39255631  2.26836661  2.24079808
#>  [199]  2.49018036  1.93415899  2.83879092  2.11303966  1.51774075  3.03877760
#>  [205]  2.59849402  0.65488715  2.72806244  1.17884939  0.17669859  2.15719871
#>  [211]  0.15514763  2.42822603  2.42822603  1.25667525  0.31329578  1.14045628
#>  [217]  0.95532579  1.63389965  2.17121808  1.29001363  1.76115996  1.25941899
#>  [223]  1.42414047  2.83953430  2.67740439  2.95166975  1.68227772  3.40071089
#>  [229]  1.62936448  0.85561182  3.01245972  0.92991875  0.77365439  1.96666960
#>  [235]  3.00024939  3.28503639  0.61310141  1.94201300  2.40353511  1.19232925
#>  [241] -0.09550114  1.23995297  2.63687470  1.19868790  1.04309578  1.62212216
#>  [247]  2.92205143  2.95577383  2.18272258  2.64482555  0.48114460  2.04582163
#>  [253]  1.87306332 -0.48062134  1.20106309  2.72936162  0.64610937  3.27270717
#>  [259]  0.28593226  1.39754699  0.18411421  2.40169555  0.81423011  1.58414644
#>  [265]  1.76375833  1.14836361  3.37076887  0.78232687  3.61412792  1.86824079
#>  [271]  0.86233170  2.10578513  1.97974267  1.79262966  1.33904094  3.48216209
#>  [277]  1.36733642  0.88696460  2.18699788  0.60802340  1.81861812 -0.33457214
#>  [283]  1.15571465  0.56439084  2.53973482  1.20095316  2.89299092  0.71214121
#>  [289]  1.78068714  2.27025387  2.06220738  1.08875116  3.34335307  1.80834886
#>  [295]  1.26597645  1.26006793  2.14407649  1.21207573 -2.75006981  3.17121785
#>  [301]  2.52956041  0.61595411  3.22833031  3.14958916  3.82252458  3.07133976
#>  [307]  1.13923521  1.03993387  3.44022105  1.73357169  3.91978792  2.90468063
#>  [313]  2.70803946  0.82956052  4.01176722  1.94481665  1.67873024  3.08522677
#>  [319]  2.73714144  3.38938308  1.78975322  1.85503315  2.68328479  1.50043339
#>  [325]  1.52681426  1.94564739  2.24680402  1.37832637  1.31035711  0.92777893
#>  [331]  1.34967920  2.23747659  1.66338801  2.54412817  1.93949832  3.84327323
#>  [337]  2.34041959  1.62665557  1.20543992  1.64058558  3.19045356  2.85175173
#>  [343]  3.27029989 -0.29061181  3.47756378  3.22449876  1.87360099  2.63433780
#>  [349]  2.34452145  1.95996065  1.92284893  0.51518892  2.13868280  1.86215339
#>  [355]  1.94798224  4.67569714  1.61282318  1.34482648  1.75862888  1.65518633
#>  [361]  1.91067522  2.04404224  3.39635454  1.87228118  1.90006205  2.13868280
#>  [367]  0.15782769  3.43364346  1.57178256  4.63798128  2.23237403  2.04628161
#>  [373]  3.46524125  2.02383227  1.80821689  2.07126869  1.63056250  2.54983769
#>  [379]  0.10987772  1.83567901  1.25431766  2.48009915  1.41915723  1.52965346
#>  [385]  2.15499145  3.70594696 -0.28194095  3.32677095  1.87966418  2.10558633
#>  [391]  3.99745324  1.55812086  1.99856187  3.89588884  3.77808186  3.42453266
#>  [397]  1.81677761  2.19658826  1.77536759  1.51930238  2.05827236  3.10030703
#>  [403]  2.92326945  2.10587800  2.29381554  0.91106401  1.79888417  2.67948613
#>  [409]  3.04804372  3.50287327  2.86365029  2.10789329  3.58978806  1.72754828
#>  [415]  1.96350768  1.82919623  3.86178071  1.35734911  1.98331961  3.55699494
#>  [421]  2.33318865  0.33378792  1.81271763  1.87268863  0.07939267  1.59116606
#>  [427]  4.67063598  5.25186747  5.80784312  2.14972593  2.28486539  2.09380743
#>  [433]  4.78504057  1.62716609  3.70046592  1.86838471  2.20729621  1.61991796
#>  [439]  1.77143505  1.61991796  2.38545261  2.21020362  3.74839551  3.40623513
#>  [445]  1.88220243  2.74557869  2.45981970  2.48980577  2.28032701  2.49807785
#>  [451]  4.23368979  2.52989391  2.59780943  2.45304004  3.75197506  2.48980577
#>  [457]  2.49824529  2.55615346  3.33953104  1.93195431  2.52501761  2.94709171
#>  [463]  3.38900481  2.74764906  2.33555991  1.96482182  4.66543938  2.92528097
#>  [469]  3.12002394  2.29980988  3.19779522  2.83128396  4.53536185  3.38698261
#>  [475]  1.76854000  3.11151338  2.79375543 -0.86677482  3.06584623  2.12109409
#>  [481]  4.20163913  3.57745841  4.83825320  4.28016048  2.17954723  2.25650264
#>  [487]  1.96389737  1.12702691  2.42451663  4.62654854  3.47534571  3.23082867
#>  [493]  2.25615596  2.35024310  1.95823140  2.43676940  0.32568673  4.12169719
#>  [499]  4.92582143  4.14635067  4.43675970  3.11677301  3.34475616  3.26580201
#>  [505]  2.98354412  5.06105114  1.46841915  5.38613437  3.75547063  4.25938458
#>  [511] -0.10976488  3.03492810  3.69930400  3.14044768  3.12086214  4.15293048
#>  [517]  2.98127075  3.99768217 -0.35398504  2.97702915  5.20639521  3.67008569
#>  [523]  4.35526891  4.34339153  4.60734460  1.02229384  4.23555332  3.49136278
#>  [529]  6.23164106 -0.73213442  6.65914329  3.98747765  5.24720195  3.52351064
#>  [535]  6.83949436  0.07215180 -1.06854684  0.16100615  0.44678247  0.43315286
#>  [541] -0.10079426  3.06433482  4.61374803  4.26642723  3.75676158  4.13011292
#>  [547]  4.05504812  5.23447514  0.54320694  0.26358850  0.47923280  1.20245467
#>  [553] -0.33440371  3.60480349  5.03799222  4.17442486  0.94032030 -0.84196436
#>  [559]  1.17200396  2.84595264  0.82206120  0.37274485  2.62865316  1.54254898
#>  [565]  0.34930257  3.35566735  2.61556407  0.72745670  3.65243665  2.46128995
#>  [571]  2.04239169  3.09122324  1.93178301  0.72266765  3.19908188  1.71153471
#>  [577]  3.23303968  2.44735657 -0.34046193  2.52795400  2.31663297  1.02046448
#>  [583]  0.30053738  0.03439801  1.44552645 -0.28297008 -0.05359628  1.38017913
#>  [589]  0.79243316  2.65741366  2.19443573  1.56331062  0.02061230  3.80736290
#>  [595]  1.43335855  1.16154812  4.33192061  0.57137921  0.87993187  4.60455368
#>  [601]  1.37004326  1.42974011  3.19568580  1.35103257  1.36075138  1.74941627
#>  [607]  0.04450224  1.68423209  1.39919257  2.87960962  1.83390217  1.66876834
#>  [613]  1.76391927  0.58754943  3.33107676  3.24056100  1.90729682  1.50235623
#>  [619]  3.37939931  1.48313360  1.52958825  2.13012720  2.03571818  2.37949339
#>  [625]  3.09800295  1.50962024  0.22642826  1.24724197  1.53492749  1.19980790
#>  [631]  1.99926854  1.88464256  2.39671056  2.92585700  1.79840483  2.68684025
#>  [637]  2.69582679  3.86711302  3.23477431  2.63882016  2.85708426  1.98060621
#>  [643]  1.82458996  2.55023124  2.81165120  3.48882850  1.61761933  3.95978480
#>  [649]  2.25949158  2.13363580  1.20988843  2.22934956  1.98518542  2.08235507
#>  [655]  2.22223289  2.57025667  1.28315518  2.57460140  3.23886008  2.08637661
#>  [661]  2.68321245  1.07853455  2.61303559  3.48569156  3.84252931  3.52689545
#>  [667]  2.62372010  2.47400812  2.55102202  2.28441657  1.78789095  3.00505167
#>  [673]  1.42512225  2.12897264  2.02097509  2.12943268  1.87190663  2.51703177
#>  [679]  1.84668372  1.62795235  2.61682582  2.07902538  1.49666236  1.78954124
#>  [685]  1.43065456  1.19936785  2.26893635  2.42806920  2.50375078  2.78842705
#>  [691]  1.67792977  2.38307478  2.11808710  2.39554277  2.97456240  2.24011090
#>  [697]  3.42408655  2.07366955  1.89735276  2.20297747  2.48659082  2.69900519
#>  [703]  2.06677043  2.21659692  2.84902153  3.58248775  3.57745650  3.58951649
#>  [709]  3.47544471  2.92680732  3.39805023  5.14890527  3.04034041  2.22302725
#>  [715]  4.38766610  5.03225521  5.29484031  1.02325019  2.69245784  2.49343203
#>  [721]  2.23769200  1.49479959  1.41587049  1.44922585  1.85351325  1.37985474
#>  [727]  1.49030317  0.83513612  3.36371068  4.23638545  4.00612353  4.21052304
#>  [733]  3.88632749  3.94353482  3.71756999  3.33760044  3.42221209  3.54601607
#>  [739]  4.22028005  4.18488502 -0.16212599  4.04014889  3.46776328  2.95417121
#>  [745]  3.76126857  3.46775817  0.51817085  2.89209189  2.80727717  3.62248373
#>  [751]  3.86104976 -0.18241765  0.93921904  3.70525847  1.63443089  2.43718262
#>  [757]  2.90832587  2.33495255  2.30508579  1.76630774  1.98572361  1.80773511
#>  [763]  0.60768891  1.24569432  2.61209995  3.52387960  2.64435353  2.06969196
#>  [769]  2.21997924  1.72627403  3.35042252  2.10889268  2.09549945  2.88778738
#>  [775]  2.88468560  3.59390191  2.45624807  3.37754538  2.62880307  2.61171299
#>  [781]  1.29638245  3.40204416  2.60158561  1.20316935  1.71267140  3.08726162
#>  [787]  3.25355282  2.28729865  1.03321808  1.13218134  2.38850702  1.18475727
#>  [793]  2.57776595  2.13003670  1.10652215  1.84150657  2.62334347  3.01581911
#>  [799]  2.90756502  1.25945797  2.70314847  2.11654158  1.83729565  2.19720590
#>  [805]  2.69223950  0.25382039  3.37072235  2.72318454  2.79339183  2.82372254
#>  [811]  3.28577166  3.25679310  3.57511360  2.54612189  3.60093209  2.25260092
#>  [817]  2.30215625  1.83709314  3.41978388  4.01353323  3.41660117  4.37468597
#>  [823]  0.29674498  3.57935835  3.70495425  3.98561637  3.45428191  3.66086337
#>  [829]  3.27497089  3.14702539  4.23455078  1.76968264  3.18460459  2.81532365
#>  [835]  3.90984128  3.90256934  4.00893968  3.04089864  3.46683945  3.50718720
#>  [841]  3.40930517  4.95646668  3.27164935  1.58848984  2.90261349  4.09175507
#>  [847]  3.81362927  3.88564651  2.81371095  4.53622110  3.78021124  4.15543848
#>  [853]  4.32904620  3.95981489  3.99602330  3.83388189  2.76998912  4.58693927
#>  [859]  4.42723853  3.97654608  4.21212374  4.68451308  1.81950652  4.10562453
#>  [865]  4.57932529  5.15001640  5.57831159  5.57807989  5.09739506  5.01455994
#>  [871]  0.10896540  0.12535700  0.06271390  0.36057485  0.43108839  0.93320995
#>  [877]  0.16181753 -1.01411812 -0.10253743  0.30790409 -0.49951206  1.59514054
#>  [883]  2.08502953  2.40905181  2.82254816  4.84046308  3.74823508  1.30450749
#>  [889]  1.01459294  1.50277127  1.62491853  1.71717767  0.36370044  2.26131572
#>  [895]  1.88923418  0.29393852  0.33754865  0.59342459  1.60689411  1.22826065
#>  [901]  1.60922010  1.66156897  1.88934216  1.51557287  2.25085324  1.27091542
#>  [907]  0.54446153  2.84840128  2.53119334  0.45189876  0.39820538  0.89873987
#>  [913]  0.61200794  0.67950270  0.81035050 -0.27761592 -0.18237907 -0.14164102
#>  [919]  0.03506333  0.30686449  0.56680784  0.91550647 -0.60697486 -0.40537849
#>  [925] -0.36208492 -0.31864564 -0.05437869  2.73532574  1.93536313  0.62171751
#>  [931]  0.98490465  3.80736290  0.53687998  1.69084326  4.65621160  4.28264442
#>  [937]  0.46110121 -0.34046193  1.50065578  2.04239169  1.72653109  2.44735657
#>  [943]  4.72436140  4.17442486  1.07618295  1.35734911  0.19316557  2.31663297
#>  [949]  0.80594103  2.14192424  1.93178301  2.78284665  1.03068374  1.01076709
#>  [955]  0.21041549  1.05094298  0.05928069  1.90677231  0.52526021  1.56331062
#>  [961]  1.12259152  1.32515959  1.17843465  2.06983829  2.14192424  0.93414463
#>  [967]  2.20041325  3.58129049  2.97850838  3.03047152  0.69043958  0.59915148
#>  [973]  3.27879775  0.80955659  0.18395801  2.54807064  2.73532574  1.15388766
#>  [979]  1.99028693  0.33342454  0.50063027  0.36747631  1.17422988  4.80446578
#>  [985]  0.82812425  0.14862912  0.26265082  0.21386016  2.53847394 -0.03859852
#>  [991]  4.17015248  2.85175173  0.34930257  2.71290613  2.21749003  1.15036197
#>  [997]  0.58697193  1.29162488  0.21005844  0.75058928