Print enpls.ad Object

Print enpls.ad object.

# S3 method for enpls.ad
print(x, ...)

Arguments

x

An object of class enpls.ad.

...

Additional parameters for print.

See also

See enpls.ad for model applicability domain evaluation with ensemble partial least squares regressions.

Examples

data("alkanes")
x = alkanes$x
y = alkanes$y

# training set
x.tr = x[1:100, ]
y.tr = y[1:100]

# two test sets
x.te = list("test.1" = x[101:150, ],
            "test.2" = x[151:207, ])
y.te = list("test.1" = y[101:150],
            "test.2" = y[151:207])

set.seed(42)
ad = enpls.ad(x.tr, y.tr, x.te, y.te,
              space = "variable", method = "mc",
              ratio = 0.9, reptimes = 50)
print(ad)

#> Model Applicability Domain Evaluation by ENPLS
#> ---
#> Absolute mean prediction error for each training set sample:
#>   [1]  0.952964269  0.227857240  0.024877237  1.163879713  0.141095062
#>   [6]  1.058032012  0.003849115  0.618109097  0.383895011  0.515935569
#>  [11]  0.192704768  0.248833459  3.538291562  0.679178994  0.176832135
#>  [16]  3.232019616  0.871518341  0.076978994  0.777746239  1.200179506
#>  [21]  0.135295007  0.321002807  0.641236588  0.152722140  0.472424711
#>  [26]  0.745998942  0.019055663  0.712422343  0.769621211  0.459522202
#>  [31]  1.731525210  0.722384508  1.355793308  0.838485422  0.842309815
#>  [36]  1.166516548  0.471435330  1.189135757  1.299007678  1.118939885
#>  [41]  2.281011276  2.310439064  1.346321413  1.208930542  1.438794633
#>  [46]  1.427529070  1.065135803  0.832663678  2.072275285  2.101168885
#>  [51]  1.514319006  0.015058809  3.819176717  1.007670697  0.439658110
#>  [56]  0.809523355  2.453945281  2.558114777  1.406495117  0.619299765
#>  [61]  0.812277377  3.432698537 14.257669298  5.914647782  0.733712575
#>  [66]  1.133440427  2.414174267  1.283247682  1.672028799  2.179676545
#>  [71]  0.741112184  1.134723134  1.078288339  2.228295304  1.487436399
#>  [76]  2.118049098  0.324518841  1.699596999  1.931549533  0.588095468
#>  [81]  0.872726501  0.498811129  5.508663134  0.439650487  2.590822060
#>  [86]  0.832603329  2.119312840  1.911250038  1.502272652  0.367572720
#>  [91]  1.512451252  3.751992767  3.295632182  3.943191803  6.394816913
#>  [96]  1.785290284  3.525612606  1.928341882  1.684967457  0.957378697
#> ---
#> Prediction error SD for each training set sample:
#>   [1] 0.6037920 1.1903722 0.5381600 0.4877270 0.6090139 0.3020451 0.5032694
#>   [8] 0.3020217 0.4341417 0.2849416 0.3082615 0.2600513 0.6880418 0.2841161
#>  [15] 0.1605102 0.5045498 0.2826224 0.2470310 0.3260170 0.3993696 0.3366345
#>  [22] 0.2377636 0.4520878 0.3670095 0.2414023 0.4950292 0.4037808 0.2453134
#>  [29] 0.5135750 0.3992922 0.3224348 0.2112565 0.5075182 0.2269638 0.2000231
#>  [36] 0.2548181 0.5697963 0.4022667 0.3217699 0.2552323 0.5292518 0.2612220
#>  [43] 0.2996296 0.7522174 0.5269289 0.5132615 0.7455050 0.4586270 0.3047196
#>  [50] 0.3243730 0.3792627 0.2875333 0.4033769 0.3080942 0.2440557 0.4226905
#>  [57] 0.3617896 0.3367685 0.4915875 0.3291164 0.3831012 0.2107170 0.4295075
#>  [64] 0.2699011 0.2777131 0.5809300 0.3207287 0.3383381 0.3742494 0.4166662
#>  [71] 0.2151759 0.3130959 0.3207774 0.2630901 0.3891006 0.4161887 0.5438654
#>  [78] 0.5195780 0.3231830 0.4442358 0.4541743 0.1906334 0.3325158 0.3493930
#>  [85] 0.2555605 0.4065463 0.1811006 0.4477221 0.2376643 0.4568802 0.4092832
#>  [92] 0.3274610 0.8199297 0.5733589 0.2484411 0.2312196 0.2973911 0.4134747
#>  [99] 0.1980512 0.4666923
#> ---
#> Absolute mean prediction error for each test set sample:
#> [[1]]
#>  [1]  1.65170701  0.49531615  1.78361645  0.23767718  4.71325994  0.43986755
#>  [7]  0.77854629  1.89417222  3.36962545  2.76639243  2.25568296  2.89836420
#> [13]  3.63100994  2.01848531  0.03109045  2.36063476  3.75040484  0.02132243
#> [19]  2.51941440 16.03584882 16.77582884 14.66194434 18.13089220 11.32336967
#> [25] 12.79055091 34.99014814 16.67366985 15.54186441 15.30161748 10.75800031
#> [31]  6.45099738 15.84129844 15.62710177 16.57239062  7.26465800 16.98152969
#> [37] 15.39361603  7.75569311  5.65622109 16.61716965  9.90217528 15.24220847
#> [43]  8.02113177  5.59807374 16.82005939 11.11522273  7.22369618  5.94469280
#> [49] 11.19291635 11.27725989
#> 
#> [[2]]
#>  [1]  6.9144580  5.6570865 15.5179131 14.1676929  4.9102502  5.7374884
#>  [7]  4.6252496 11.3201641  4.5588885  3.0633023 48.0599221 41.0910309
#> [13] 44.8303046 42.1314954 45.0991043 39.9185727 50.1783712 44.2194596
#> [19] 33.0747136 57.8330979 38.3364732 32.6318155 35.7610257 44.6608655
#> [25] 40.2281801 31.9055346 29.2181359  0.1609417  1.1632331  3.3782633
#> [31]  2.6223788  2.0729561  1.6876716  0.6114954  6.3320391  1.4802177
#> [37]  5.2190920 16.2389199  0.7500703  1.4456360  3.8394186  0.9919153
#> [43]  4.9036551  0.8574751  4.3191447 20.3702347  7.3236414 17.5135789
#> [49]  6.2939005  9.8342026 15.5669774 14.3008963 15.0035390 15.8457797
#> [55] 10.7864861 16.7106656 14.6654083
#> 
#> ---
#> Prediction error SD for each test set sample:
#> [[1]]
#>  [1]  0.3393848  0.4929920  0.3107950  0.4184266  0.6669721  0.8136573
#>  [7]  0.3005431  0.5311227  0.4758082  0.6285845  0.5000884  0.6686000
#> [13]  0.5725195  0.6659227  0.6516754  0.8552208  0.7149362  0.8151977
#> [19]  0.9950084  9.7517793  9.4627176  9.1793272  9.1616135  7.7062602
#> [25]  6.7459009 97.4870122  8.2650312  8.3206868  7.9405210  8.3188809
#> [31]  8.2031790  7.3603210  7.7438013  8.0265727  8.0815870  8.0793945
#> [37]  7.6833849  7.9394597  7.9554483  7.1803562  7.8552396  7.8081934
#> [43]  7.8941798  7.8886605  7.8845899  7.7425063  7.7395051  7.7539773
#> [49]  7.8556897  7.6312828
#> 
#> [[2]]
#>  [1]   7.7262583   7.7000653   7.7620703   7.3616336   8.1015145   7.6048302
#>  [7]   7.5446912   6.6314803   7.5014084   7.5212929  24.1233998  20.8707767
#> [13]  20.8872972  22.3588492  21.4742666  21.7505039  21.9294556  21.8516360
#> [19]  21.8039440  84.7711961  21.6591256  21.6215360  21.1477441  21.3507527
#> [25]  21.4009547  21.2622910  21.1056920   0.3187686   0.6723836   0.8939262
#> [31]   0.5571608   0.6040339   0.3766123   0.4162408   0.5763224   0.1392987
#> [37]   0.3558719 102.9989191   0.2394275   0.3607086   0.2912192   0.1952102
#> [43]   0.3351914   0.3785477   0.7429789   9.4156095   7.9078708   8.5296029
#> [49]   8.1904502   8.2091738   8.2054809   7.0927898   8.0714083   7.9687650
#> [55]   8.0388086   8.1221421   7.8760553
#>