Print enpls.fs Object

Print enpls.fs object.

# S3 method for enpls.fs
print(x, sort = TRUE, nvar = NULL, ...)

Arguments

x

An object of class enpls.fs.

sort

Should the variables be sorted in decreasing order of importance?

nvar

Number of top variables to show. Ignored if sort = FALSE.

...

Additional parameters for print.

See also

See enpls.fs for measuring feature importance with ensemble partial least squares regressions.

Examples

data("alkanes")
x = alkanes$x
y = alkanes$y

set.seed(42)
fs = enpls.fs(x, y, reptimes = 100)
print(fs)

#> Variable Importance by Ensemble Partial Least Squares
#> ---
#>          Importance
#> MEDV.33  2.32584140
#> Chi.P.4  2.01574084
#> MEDV.23  1.59626805
#> Estate.1 1.37064152
#> Chi.C.3  1.13547497
#> MEDV.22  1.04085558
#> Chi.P.5  0.91240637
#> MEDV.12  0.76218891
#> Estate.2 0.74572241
#> Chi.P.3  0.69494972
#> Kappa.1  0.69034300
#> MEDV.11  0.64388018
#> Kappa.3  0.44209890
#> Chi.PC.4 0.37039029
#> MEDV.13  0.24963703
#> Chi.P.2  0.24904390
#> Estate.3 0.22900656
#> Chi.P.0  0.10418293
#> Kappa.4  0.08937748
#> Chi.P.1  0.08733047
#> Kappa.2  0.08014690

print(fs, nvar = 10L)

#> Variable Importance by Ensemble Partial Least Squares
#> ---
#>          Importance
#> MEDV.33   2.3258414
#> Chi.P.4   2.0157408
#> MEDV.23   1.5962680
#> Estate.1  1.3706415
#> Chi.C.3   1.1354750
#> MEDV.22   1.0408556
#> Chi.P.5   0.9124064
#> MEDV.12   0.7621889
#> Estate.2  0.7457224
#> Chi.P.3   0.6949497