Ensemble Partial Least Squares Regression

Ensemble partial least squares regression.

enpls.fit(x, y, maxcomp = NULL, cvfolds = 5L, reptimes = 500L,
  method = c("mc", "boot"), ratio = 0.8, parallel = 1L)

Arguments

x
Predictor matrix.

y
Response vector.

maxcomp
Maximum number of components included within each model. If not specified, will use the maximum number possible (considering cross-validation and special cases where n is smaller than p).

cvfolds
Number of cross-validation folds used in each model for automatic parameter selection, default is 5.

reptimes
Number of models to build with Monte-Carlo resampling or bootstrapping.

method
Resampling method. "mc" (Monte-Carlo resampling) or "boot" (bootstrapping). Default is "mc".

ratio
Sampling ratio used when method = "mc".

parallel
Integer. Number of CPU cores to use. Default is 1 (not parallelized).

Value

A list containing all partial least squares model objects.

See also

Examples

data("alkanes")
x = alkanes$x
y = alkanes$y

set.seed(42)
fit = enpls.fit(x, y, reptimes = 50)
print(fit)

#> Coefficients of the Models by Ensemble Partial Least Squares
#> ---
#>                      [,1]         [,2]          [,3]          [,4]
#> (Intercept)    -0.1341459   -0.2549992     0.1216662  8.535263e-03
#> Chi.P.0      -624.4054630  128.0898938  -384.3587022  1.028365e+02
#> Chi.P.1      -711.1961530  145.4635882  -624.6777603  1.155413e+02
#> Chi.P.2      2793.5273367   73.6684221  2518.6115032  8.346000e+01
#> Chi.P.3      1636.2871339  364.0672778  1655.9250485  3.099410e+02
#> Chi.P.4        -3.3281156   -2.4427978    -2.8047048 -7.086803e+00
#> Chi.P.5       327.0449131  184.7246535   331.4697201  9.020049e+01
#> Chi.C.3        44.7393444   48.3609276    42.6496921  1.687598e+01
#> Chi.PC.4     -166.0811461   48.0608478  -163.6200094 -3.242323e+01
#> Kappa.1      2763.7158330 -172.8878682  2827.5697820 -5.293024e+01
#> Kappa.2      -767.1078266  199.2050049  -634.1651364  1.356848e+02
#> Kappa.3     -2495.5625843   43.9046324 -2473.8883592  1.057668e+02
#> Kappa.4      -409.6499447  134.1205625  -416.1474374  1.096061e+02
#> Estate.1      -18.3771676 -142.7997304   -81.4318449 -1.166073e+02
#> Estate.2      -49.6204561   75.3891887    -6.3089949  8.792373e+01
#> Estate.3        8.8368070    5.0178780    11.9322457 -2.520637e+00
#> MEDV.11      -408.3768796  -28.9692956  -426.3080329 -3.674518e+01
#> MEDV.12     -1386.2334467 -224.4159924 -1409.1733790 -1.902066e+02
#> MEDV.13        31.7539378  -13.9717807    27.3830380  2.147782e+01
#> MEDV.22      -939.3931805 -439.7581385 -1060.5178999 -2.260530e+02
#> MEDV.23       -47.4006311  -12.4775404   -43.0369210 -3.580110e+01
#> MEDV.33        -8.2599846   -1.6681701    -7.8627385 -3.849338e+00
#>                      [,5]         [,6]         [,7]          [,8]         [,9]
#> (Intercept)    0.07598922    0.1065228    0.1458699    0.09264789    0.1490357
#> Chi.P.0       85.54139701  140.5300740   63.9098771   84.79566097   73.7648681
#> Chi.P.1       86.57424842  151.9635477   64.6083160   84.42529173   73.9890196
#> Chi.P.2       87.10051533  148.1512863   62.6656563   96.11784246   75.1185747
#> Chi.P.3      110.67657490  368.5493646   73.3703484   83.47085817  100.2949146
#> Chi.P.4      -10.33652761   -4.7756756  -10.8745475  -10.29947672  -13.1537267
#> Chi.P.5       67.13847673   89.4579446  102.2165823   37.96046456   59.1034556
#> Chi.C.3       20.60231038   15.4076117   35.3315895   17.29979087   21.4296941
#> Chi.PC.4      13.89133986  -45.0002036   53.0209234   -8.27363877    8.8460415
#> Kappa.1        0.60790775 -111.6924989    2.5263723   24.70472022    7.8288888
#> Kappa.2       93.60776359  188.0784560   70.4254707   85.33696400   75.3401284
#> Kappa.3       85.73127647  134.7715700   58.7101915  102.53643459   91.8175085
#> Kappa.4       86.57427992  146.1307505   64.5425522   85.43962813   74.4051803
#> Estate.1    -144.80793746 -137.5748197   -5.1793248 -127.59679563 -118.0492887
#> Estate.2     116.09805890  122.0227719   21.5472859   91.38417806   91.0059542
#> Estate.3     -14.75290315    5.6039686  -17.6987623  -14.50755370  -14.1700364
#> MEDV.11      -31.81786666  -44.4560028  -25.8387949  -34.34055109  -36.0932124
#> MEDV.12     -121.00624656 -244.7632257 -101.8503806 -120.46798229 -116.8289065
#> MEDV.13       38.60746824   33.0869182    4.4031446   47.81071206   35.5611745
#> MEDV.22       38.46024786 -448.3271380   98.7572478   74.30396277   44.6395879
#> MEDV.23      -40.07087316  -58.2829913  -12.5184050  -45.75453092  -41.0492258
#> MEDV.33       -5.00656323   -5.3301899   -1.8495780   -5.17544833   -4.7531883
#>                     [,10]         [,11]        [,12]       [,13]        [,14]
#> (Intercept)     0.2187588  9.986015e-02    0.1220749   -0.127940    0.1280691
#> Chi.P.0      -202.1013422 -4.480129e+02   62.4793762  123.853086  129.6467670
#> Chi.P.1      -333.5650621 -8.010975e+02   62.7404894  145.454424  137.4486139
#> Chi.P.2      1552.9688140  2.830766e+03   62.7712449   54.047391  137.4227777
#> Chi.P.3      1078.4361563  1.814958e+03   70.7371881  470.237976  365.4244554
#> Chi.P.4        -8.2855341 -7.304622e+00  -12.5218970   -4.977245   -5.4079438
#> Chi.P.5       219.8074868  3.477547e+02   90.2637889  191.241147  107.8145596
#> Chi.C.3        25.2032517  4.169297e+01   30.0862619   41.446875   21.4697991
#> Chi.PC.4     -111.4793561 -1.864561e+02   49.1921196   16.985096  -25.0364182
#> Kappa.1      1625.1576971  2.471031e+03    6.3872684 -200.430596  -89.8023741
#> Kappa.2      -302.4701590 -7.630739e+02   67.1298801  199.539471  174.2558514
#> Kappa.3     -1395.4238430 -2.273700e+03   65.3631475   64.864077   75.9446069
#> Kappa.4      -225.6182339 -4.814663e+02   62.9554288  133.088130  134.0053017
#> Estate.1      -71.7979260 -8.304568e+01    7.3760146 -161.357983 -167.7137462
#> Estate.2       42.1518687  2.465991e+01   16.0447298  115.555398  137.1836231
#> Estate.3        6.2908344  1.755058e+01  -15.2606997    9.998935    5.3325852
#> MEDV.11      -260.0705355 -3.888007e+02  -23.2578081  -19.604650  -42.1389749
#> MEDV.12      -920.6602962 -1.368257e+03 -106.8679166 -221.721077 -240.3240369
#> MEDV.13        27.4139872  3.720815e+01    3.0571460  -14.204065   27.6720008
#> MEDV.22      -705.1002573 -1.140436e+03  100.7088744 -516.896451 -431.8837373
#> MEDV.23       -36.6326527 -3.882574e+01   -7.6713893   -2.819987  -43.8073620
#> MEDV.33        -6.1249354 -8.398871e+00   -1.7345765   -1.939215   -4.8052452
#>                     [,15]        [,16]         [,17]         [,18]
#> (Intercept)    0.05322357    0.0499378     0.2093652     0.1500331
#> Chi.P.0       62.11199076   83.0549431  -472.6872786  -619.3704716
#> Chi.P.1       62.60087895   84.2171792  -668.0620821 -1001.7258725
#> Chi.P.2       61.26822165   82.3936600  2803.1273111  3383.1327165
#> Chi.P.3       74.56588465  114.6737983  1787.0738944  1821.5446974
#> Chi.P.4      -13.70404249  -10.2221472    -1.6705986    -6.4277922
#> Chi.P.5       85.73980080   64.2134842   280.0952892   317.8561953
#> Chi.C.3       28.98301333   19.2611807    21.1296477    34.5214101
#> Chi.PC.4      44.89460309   11.1676923  -272.3110009  -212.9152433
#> Kappa.1        3.65134689    3.3702615  4393.3486078  2438.8348015
#> Kappa.2       67.13326250   91.0634076  -868.4359923  -950.8442609
#> Kappa.3       61.17523501   78.4792705 -3534.5654919 -2342.3511277
#> Kappa.4       62.68724379   84.0947255  -334.9987264  -554.4368119
#> Estate.1     -10.41101342 -137.1430717    81.1981148   -12.7733158
#> Estate.2       8.87463405   96.1114137  -118.8006974   -21.6051297
#> Estate.3     -14.52206478  -17.2519826     8.0131320    11.7818868
#> MEDV.11      -22.11574988  -27.5626251  -625.8252282  -391.7865558
#> MEDV.12     -103.20589034 -109.7073001 -1997.8884765 -1326.5465782
#> MEDV.13        0.88394212   30.2121046    44.0007744    66.1939257
#> MEDV.22       99.88974094   61.8694526 -1066.3092681  -994.4199076
#> MEDV.23       -2.50616987  -27.8772743   -91.5702972   -62.4624039
#> MEDV.33       -1.23611018   -3.9084750   -11.3075720    -9.6378891
#>                     [,19]         [,20]         [,21]         [,22]
#> (Intercept)    -0.1502886    0.08036824     0.0249751     0.2088137
#> Chi.P.0      -735.6992313   78.73004700  -567.9559160 -5104.4326459
#> Chi.P.1      -968.4335945   80.02661497  -769.0118386 -6075.6637444
#> Chi.P.2      3492.2810385   78.28159293  2993.7589298 16698.2074751
#> Chi.P.3      2148.3521727  113.47458156  1913.1710147  3957.0289596
#> Chi.P.4        -2.5888151  -10.12752672    -4.3002039    -9.4746485
#> Chi.P.5       365.1218310   58.32496429   362.2692270   766.4197898
#> Chi.C.3        35.1710551   15.15570913    45.7212197   105.3151164
#> Chi.PC.4     -280.4723606    5.65589017  -213.2826724   -22.3216546
#> Kappa.1      3747.4929012    9.32922509  3549.0162266 -6370.4894877
#> Kappa.2     -1011.5459256   83.58308762  -816.7302487 -2479.4786048
#> Kappa.3     -3239.9325873   87.29794350 -3022.5626284  1723.5507906
#> Kappa.4      -557.1602765   79.84525313  -537.1752045 -5975.0665661
#> Estate.1       26.6067289  -98.50958486     6.5929136  -137.0457779
#> Estate.2      -75.0650488   60.79373279   -65.1670690    28.9792051
#> Estate.3       14.0265018  -13.79173107    11.5909252    -9.0817955
#> MEDV.11      -525.6516654  -29.31947991  -490.9215890   600.3875698
#> MEDV.12     -1793.3535506 -116.29956518 -1669.1102237  1701.5665165
#> MEDV.13        46.8541866   25.87717419    27.4979804   450.4102597
#> MEDV.22     -1230.2828381   64.51887624 -1174.1396358   414.0654654
#> MEDV.23       -70.3555084  -27.02169863   -53.2693549    26.6927815
#> MEDV.33       -10.6874696   -3.60547713    -8.8599342    -7.3699438
#>                     [,23]        [,24]      [,25]         [,26]         [,27]
#> (Intercept) -1.664861e-01    0.3296991   0.100649 -5.747179e-02   -0.06821886
#> Chi.P.0      5.244973e+04   60.0547725  58.852901 -2.514666e+02  129.56218816
#> Chi.P.1     -3.520574e+04   60.2314507  59.112182 -3.680877e+02  141.64543299
#> Chi.P.2     -2.065928e+03   60.3323259  58.584207  1.652507e+03  146.39601879
#> Chi.P.3      7.718817e+02   69.7470849  68.007596  1.068592e+03  343.86530003
#> Chi.P.4     -9.456504e+00  -14.0809056 -11.268615 -6.387670e+00   -4.03853622
#> Chi.P.5      2.078462e+02   83.0836268  87.037192  2.402042e+02  121.73499672
#> Chi.C.3      3.124237e+01   30.0799062  30.136529  3.716873e+01   25.21576713
#> Chi.PC.4    -1.283457e+02   48.4822197  47.198181 -8.751519e+01   -7.33499282
#> Kappa.1     -1.460035e+04    4.6581815   6.148191  1.423926e+03  -79.93015117
#> Kappa.2     -5.692755e+04   63.7638890  62.851585 -3.586798e+02  181.51306890
#> Kappa.3     -7.297744e+03   60.6077840  60.962187 -1.316210e+03   62.08015541
#> Kappa.4      5.439385e+04   60.4943111  59.296879 -2.398336e+02  139.77544631
#> Estate.1    -8.314986e+01   -7.1161286 -10.218256 -8.844007e+01 -158.48453732
#> Estate.2     5.552992e+00   10.2962068   1.615970  4.261177e+01  107.99445172
#> Estate.3     9.470736e+00  -13.7837262 -12.586731  5.720017e+00    2.44279742
#> MEDV.11     -1.974012e+02  -23.9132254 -25.126954 -2.269280e+02  -45.84970397
#> MEDV.12     -7.744293e+02 -102.7324491 -99.874553 -8.391600e+02 -249.45227420
#> MEDV.13      9.009398e+01    0.7961830   4.310260  2.776071e+01   21.16673104
#> MEDV.22     -7.167383e+02   95.3094454  99.348782 -6.901981e+02 -406.07752403
#> MEDV.23     -2.143044e+01   -4.6186826 -12.786547 -3.517220e+01  -45.79471524
#> MEDV.33     -6.980064e+00   -1.2493984  -1.812144 -6.081663e+00   -4.58495615
#>                     [,28]       [,29]        [,30]         [,31]         [,32]
#> (Intercept)     0.1073358   -0.039392   -0.1466792    -0.2525722  4.591005e-03
#> Chi.P.0      -219.8674832  143.777894  121.7879885  -392.5827103  8.401076e+01
#> Chi.P.1      -415.7581757  157.216534  132.7891057  -480.5868826  8.525340e+01
#> Chi.P.2      1781.5588315  166.919500   92.5126452  2171.7318619  8.285908e+01
#> Chi.P.3      1107.8691334  255.571906  393.2447281  1565.9484121  1.205856e+02
#> Chi.P.4        -4.2457810  -13.340543   -4.8846990    -0.7654113 -9.625364e+00
#> Chi.P.5       216.1936276   72.723121  150.1362663   334.8409837  6.266412e+01
#> Chi.C.3        23.6728339   11.930094   30.9720117    48.0241483  2.016360e+01
#> Chi.PC.4     -122.4489737  -21.042376   -0.3268302  -166.6440054  7.566169e+00
#> Kappa.1      1169.9437298  -64.906733 -123.2960705  3235.1917389  2.794415e+00
#> Kappa.2      -387.5101018  194.799107  173.8375377  -647.8182738  9.139647e+01
#> Kappa.3     -1126.6003671  105.741204   59.5549101 -2746.5969551  7.807440e+01
#> Kappa.4      -239.5322344  152.288067  128.7935768  -238.8702892  8.521083e+01
#> Estate.1     -102.8749115  -73.708002 -128.7783348  -100.0984408 -1.768602e+02
#> Estate.2       64.0395094   57.577218  106.3744254    -1.7229840  1.345669e+02
#> Estate.3        8.9947451    4.213791    7.8381966     7.0151938 -1.553030e+01
#> MEDV.11      -220.6176788  -58.159154  -33.5135423  -452.1966268 -2.812434e+01
#> MEDV.12      -794.8763838 -271.013801 -238.6576946 -1511.5550395 -1.112620e+02
#> MEDV.13        51.2547897   24.855847    3.8027226    18.7427541  4.126864e+01
#> MEDV.22      -734.4301913 -432.739123 -452.3080845  -957.7539087  4.648187e+01
#> MEDV.23       -59.3651586  -59.393045  -21.6720287   -49.4227232 -3.719982e+01
#> MEDV.33        -7.8288077   -4.963071   -3.2335001    -8.7811446 -4.617540e+00
#>                     [,33]       [,34]        [,35]        [,36]         [,37]
#> (Intercept)   -0.03217449  -0.1412315    0.1092735   -0.1880724    -0.0261019
#> Chi.P.0       87.55335562  61.0538192   62.0646441   22.0299735  -522.6360851
#> Chi.P.1       89.26288151  61.6897823   62.3085919  -12.2938434 -1068.6966734
#> Chi.P.2       81.25208417  59.9099921   61.2902223  538.6903941  3608.9693902
#> Chi.P.3      145.30623553  71.6386766   74.3483724  605.6254142  1820.9398598
#> Chi.P.4      -10.85079008 -11.8689922  -12.9826019   -4.8005378    -7.6175310
#> Chi.P.5       54.91991622  95.1621969   88.2360457  225.4121741   318.1731063
#> Chi.C.3       13.28874262  28.1759810   28.8458146   48.2304384    34.8583770
#> Chi.PC.4      -6.83424870  51.4976709   48.6933966   15.3656485  -206.6790047
#> Kappa.1       -1.39835123   4.3084808    6.6296565  460.8591720  2323.5016843
#> Kappa.2       96.61333585  66.9956804   65.9189374   10.7393686  -961.1497158
#> Kappa.3       82.97734471  60.5794307   64.1338989 -491.5473681 -2345.3277687
#> Kappa.4       88.62020936  61.6904091   62.5129808   26.0524291  -675.8638406
#> Estate.1    -156.00353954  -3.2358925   -6.5233679 -167.1599060    30.5134027
#> Estate.2     101.81914100   8.0793644    1.6477457   91.0119102   -62.1828607
#> Estate.3     -15.08002386 -18.1488031  -12.8844562    3.3164859    16.3894883
#> MEDV.11      -23.57386972 -26.1191846  -24.7336323 -101.0792882  -385.4480384
#> MEDV.12     -113.01902398 -99.7891831 -104.2897481 -441.4928033 -1337.1617350
#> MEDV.13       32.55772993   2.8086805    1.4391486   -4.9976157    66.8602539
#> MEDV.22       35.61459146  96.9693671  103.4305412 -505.6819998 -1071.6742201
#> MEDV.23      -24.31195785 -12.6806289   -9.1364769   -6.8690879   -65.9225515
#> MEDV.33       -4.19125679  -1.7627662   -1.3034020   -2.8236015    -9.1203903
#>                     [,38]         [,39]        [,40]         [,41]
#> (Intercept)  7.976432e-02   -0.09812946    0.0724689    0.01744245
#> Chi.P.0     -5.650972e+02  121.01581409   93.3944627   63.14232138
#> Chi.P.1     -7.019952e+02  132.75962655   95.1317816   64.00352667
#> Chi.P.2      2.720756e+03   97.69570488   99.3982915   61.23981940
#> Chi.P.3      1.534512e+03  379.60359104  107.0525820   77.96132631
#> Chi.P.4     -6.438292e+00   -4.29576216   -7.3689764  -10.51451030
#> Chi.P.5      3.087338e+02  130.17772308   72.5949092   82.76714838
#> Chi.C.3      4.120816e+01   28.03981487   25.8744682   25.67131852
#> Chi.PC.4    -1.421984e+02  -13.74551590   22.4995478   47.07788827
#> Kappa.1      1.960252e+03 -100.68040420   -5.5979658    0.84082457
#> Kappa.2     -6.614366e+02  169.07215363  106.9716470   69.63357675
#> Kappa.3     -1.885593e+03   75.36163710   74.3103742   56.18771779
#> Kappa.4     -4.611732e+02  128.01930709   95.0813634   63.86733909
#> Estate.1    -9.453610e+01 -160.12993264 -160.4567102   -8.21869335
#> Estate.2     2.014665e+01  120.73801919  115.8771295    3.76226832
#> Estate.3     9.885487e+00    2.57402778  -16.5256796  -15.58573868
#> MEDV.11     -3.156891e+02  -36.62033400  -31.8840722  -22.70794614
#> MEDV.12     -1.105427e+03 -226.69466696 -124.7464813 -100.47446351
#> MEDV.13      4.611872e+01   17.07755720   38.2570734   -1.53128638
#> MEDV.22     -9.149812e+02 -377.97131471   24.5707486  101.46154100
#> MEDV.23     -3.950504e+01  -35.76272244  -39.7375002   -2.27900073
#> MEDV.33     -7.381425e+00   -4.41808561   -4.5188624   -0.75321303
#>                     [,42]         [,43]         [,44]         [,45]
#> (Intercept)    0.05312986    0.09661002    0.35353908   -0.05081583
#> Chi.P.0      112.71709568   93.27574141  145.38984713   96.55765294
#> Chi.P.1      122.74456998   94.50595398  156.73074417  114.12127739
#> Chi.P.2      126.40452951   91.43117982  121.91686331   50.59360966
#> Chi.P.3      372.03942651  140.22742617  412.37377554  356.73769614
#> Chi.P.4       -9.60475347   -5.74283011   -4.00533926   -8.03933862
#> Chi.P.5      144.77456510   43.81556028  149.14326768  140.34895668
#> Chi.C.3       29.36100691   15.79899990   29.41997211   32.97715020
#> Chi.PC.4       5.10812204  -22.45245874   -5.11904803    0.79079143
#> Kappa.1      -85.55041764    8.24410527 -135.92658414  -93.96663883
#> Kappa.2      161.85607071   98.00524733  207.11905052  142.98160481
#> Kappa.3       34.11060194  100.54311833   91.92861549   62.25314691
#> Kappa.4      119.00785784   94.34129401  151.75757971  105.51827226
#> Estate.1    -129.29386907 -208.30728798  -70.40117977 -150.01843268
#> Estate.2      99.44265142  139.24069313   60.41788411   75.79964728
#> Estate.3       4.68893755  -17.10364100   10.97605619    2.16496247
#> MEDV.11      -31.42745845  -33.48338789  -53.16588028  -31.55394884
#> MEDV.12     -231.97545363 -112.22535017 -291.28547646 -186.35964442
#> MEDV.13        3.11577763   53.43702166    0.06648441   -3.58228679
#> MEDV.22     -394.29652448   45.39029099 -564.22574472 -284.55949529
#> MEDV.23      -14.51864783  -52.80266999  -44.39938000  -16.72409927
#> MEDV.33       -2.94242339   -6.19150348   -3.07928911   -2.46166826
#>                     [,46]         [,47]         [,48]         [,49]
#> (Intercept)    0.02763733 -2.949276e-03    0.00933163 -2.925492e-02
#> Chi.P.0      119.76083447  1.129441e+02    4.00605588 -1.282583e+02
#> Chi.P.1      128.00879962  1.203935e+02 -144.44015726 -7.607541e+01
#> Chi.P.2       82.64439385  1.703624e+02  873.51209256  1.198814e+03
#> Chi.P.3      414.85518481  3.804383e+02  724.62123680  9.275348e+02
#> Chi.P.4       -6.00591844 -6.789399e+00   -2.10084499 -2.235148e+00
#> Chi.P.5      136.42014791  9.664748e+01  196.39507405  2.246663e+02
#> Chi.C.3       30.27762598  1.450722e+01   36.82624995  3.708358e+01
#> Chi.PC.4     -29.91508593 -3.889861e+01  -49.60080362 -6.086177e+01
#> Kappa.1     -102.29900215 -5.799182e+01  795.57974800  1.162280e+03
#> Kappa.2      158.33308237  1.526997e+02 -123.10490658 -1.616326e+02
#> Kappa.3      106.85530971  6.713178e+01 -727.02628268 -1.070253e+03
#> Kappa.4      123.48968624  1.233223e+02  -88.59064626 -6.770034e+01
#> Estate.1    -160.96400599 -1.454809e+02 -132.26907704 -7.975430e+01
#> Estate.2     126.45306491  1.286394e+02   62.23643298  2.976367e+01
#> Estate.3       3.91299173  4.809438e+00    2.37616353  1.341223e+01
#> MEDV.11      -44.82801283 -4.370907e+01 -156.95811989 -2.106624e+02
#> MEDV.12     -235.40724062 -2.484295e+02 -562.29811222 -8.019890e+02
#> MEDV.13       19.62294015  2.907318e+01   19.71994823  1.153857e+01
#> MEDV.22     -391.42798989 -4.071422e+02 -524.09488973 -8.327103e+02
#> MEDV.23      -46.00040372 -4.396085e+01  -44.01132959 -3.827794e+01
#> MEDV.33       -4.69111047 -5.147478e+00   -5.52423087 -5.048511e+00
#>                     [,50]
#> (Intercept) -9.603262e-02
#> Chi.P.0     -9.309058e+02
#> Chi.P.1     -1.592379e+03
#> Chi.P.2      4.847776e+03
#> Chi.P.3      2.576757e+03
#> Chi.P.4     -7.036381e+00
#> Chi.P.5      4.581942e+02
#> Chi.C.3      4.882978e+01
#> Chi.PC.4    -2.830334e+02
#> Kappa.1      3.613006e+03
#> Kappa.2     -1.306353e+03
#> Kappa.3     -3.389648e+03
#> Kappa.4     -1.113008e+03
#> Estate.1    -4.680731e+01
#> Estate.2    -4.639402e+01
#> Estate.3     1.033928e+01
#> MEDV.11     -5.147027e+02
#> MEDV.12     -1.760567e+03
#> MEDV.13      7.876948e+01
#> MEDV.22     -1.162051e+03
#> MEDV.23     -5.479152e+01
#> MEDV.33     -1.101719e+01

predict(fit, newx = x)

#>   [1] -2126.0609891 -2111.9760982 -1927.2680478 -1915.4834604 -1727.4953228
#>   [6] -1717.0327322 -1527.3706534 -1517.4882170 -1327.0913476 -1317.4428952
#>  [11] -1126.6515673 -1117.1837324 -1152.3095526  -926.2769815  -917.1690948
#>  [16]  -954.0791667  -726.0468765  -717.0087604  -755.4978817  -525.9267007
#>  [21]  -517.1251877  -556.7217463  -326.0424637  -317.1951260  -357.9870566
#>  [26]  -126.2806841  -117.5483709  -159.2586266    73.3927351    82.0432158
#>  [31]    68.4537932    59.0172801    52.4317837    48.9652595    40.2208348
#>  [36]    39.5360807   272.9583143   281.5865015   267.9847379   258.5017962
#>  [41]   251.8788178   239.3896677   238.3164551   472.3781636   481.0089717
#>  [46]   436.9952328  -682.7558889  -607.9941091  -484.8138060  -484.4200700
#>  [51]  -508.1912496  -505.2121265  -520.3548845  -385.8454104  -397.3865366
#>  [56]  -408.2606489  -416.3046216  -420.4887546  -281.5743317  -286.1265441
#>  [61]  -309.0109087  -308.0712703  -301.1437542  -320.1827785  -186.0263866
#>  [66]  -195.0794268  -198.6680278  -209.5985943  -220.0159410   -81.9383083
#>  [71]   -87.1948030  -110.3977140  -108.6988338  -120.0074580    13.7178626
#>  [76]     4.4575488     2.6862410    18.1208062     0.7199899   -17.7138570
#>  [81]   117.6137951   111.9642601   111.4670223    88.7275610    88.4663316
#>  [86]    77.1412208    73.1433186   207.4716365   201.1747587   181.6169726
#>  [91]   173.4092820   174.8512612   165.3958237   314.5507535   310.4727727
#>  [96]   287.1840173   287.5285373   277.0744875   272.3217482   403.4114640
#> [101]   396.4380906   380.8697818   372.5395457   369.4481822   364.1602973
#> [106]   516.4497417   509.4947978   488.3983648   486.0758029   475.8629422
#> [111]   471.5513248   600.6951650   585.6870900   567.0510581   708.5412240
#> [116]   687.7437759   684.8524051   666.7148381   784.5338446  -567.9780122
#> [121]  -479.6045382  -369.3836113  -253.1981780  -170.9742087   -55.0224205
#> [126]   -83.6937203     8.7920940   145.2063514   114.5692364   104.2110314
#> [131]    94.6428506   229.3275463   223.5279681   206.7254723   192.6221458
#> [136]   345.1672370   312.9471092   302.1271417   292.2340349   428.3489347
#> [141]   427.0239638   404.9995619   397.4086566   390.0973710   544.2088816
#> [146]   514.6452133   500.9582264   488.5778386   491.7942231   626.9691637
#> [151]   596.1340335   586.8950902   743.2038306   710.2627474   701.9240235
#> [156]   685.7212469   784.2547577   908.6575655   883.3725020   982.6120839
#> [161]   -28.4504616   168.8803448   152.3303010   368.7605056   351.5371306
#> [166]   567.7040671   524.9811888   518.3345219   518.1022913   631.9788960
#> [171]   620.3321375   615.5170896   766.5912586   723.1935350   716.1660034
#> [176]   714.9104912   813.1544504  -340.2775074  -150.8646445  -917.3539264
#> [181]  -683.8322125  -481.2779799  -508.6674423  -398.0526307  -415.1416451
#> [186]  -309.1738199  -321.2840883  -328.0863770  -197.8107065  -109.7472644
#> [191]    79.1854570    74.3211788   203.9703959   401.4191942   380.6394837
#> [196]  -251.6804934  -166.2040459  -100.5562752   116.5113631   103.4091032
#> [201]    97.2570916   320.2850351   302.5373780   295.4997425   406.3532749
#> [206]   405.8638400   392.4606274