Ensemble Sparse Partial Least Squares Regression

Ensemble sparse partial least squares regression.

enspls.fit(x, y, maxcomp = 5L, cvfolds = 5L, alpha = seq(0.2, 0.8, 0.2),
  reptimes = 500L, method = c("mc", "boot"), ratio = 0.8, parallel = 1L)

Arguments

x
Predictor matrix.

y
Response vector.

maxcomp
Maximum number of components included within each model. If not specified, will use 5 by default.

cvfolds
Number of cross-validation folds used in each model for automatic parameter selection, default is 5.

alpha
Parameter (grid) controlling sparsity of the model. If not specified, default is seq(0.2, 0.8, 0.2).

reptimes
Number of models to build with Monte-Carlo resampling or bootstrapping.

method
Resampling method. "mc" (Monte-Carlo resampling) or "boot" (bootstrapping). Default is "mc".

ratio
Sampling ratio used when method = "mc".

parallel
Integer. Number of CPU cores to use. Default is 1 (not parallelized).

Value

A list containing all sparse partial least squares model objects.

See also

Examples

data("logd1k")
x = logd1k$x
y = logd1k$y

set.seed(42)
fit = enspls.fit(
  x, y, reptimes = 5, maxcomp = 3,
  alpha = c(0.3, 0.6, 0.9))
print(fit)

#> Coefficients of the Models by Ensemble Sparse Partial Least Squares
#> ---
#>                             [,1]          [,2]          [,3]         [,4]
#> BalabanJ            -0.061134153 -0.0319813867 -0.0503701773 -0.060698172
#> BertzCT             -0.014214644 -0.0289791361 -0.0222782338 -0.023958993
#> Chi0                 0.031019938  0.0321648253  0.0294166279  0.032023405
#> Chi0n                0.052658300  0.0544752283  0.0534163896  0.055330237
#> Chi0v                0.055520288  0.0554921440  0.0539395759  0.057338019
#> Chi1                 0.048718841  0.0487588640  0.0486358339  0.048106788
#> Chi1n                0.070330489  0.0733538898  0.0724148126  0.072450111
#> Chi1v                0.048502397  0.0476778849  0.0510553735  0.049850929
#> Chi2n                0.050882782  0.0514541495  0.0498114493  0.052094502
#> Chi2v                0.000844060 -0.0009919821  0.0015180020  0.003571898
#> Chi3n                0.039258175  0.0429630692  0.0381739239  0.037040916
#> Chi3v               -0.021377426 -0.0206986933 -0.0195837210 -0.019469855
#> Chi4n                0.031693793  0.0376532680  0.0356768615  0.029888427
#> Chi4v               -0.026517661 -0.0248985128 -0.0239647643 -0.024182098
#> EState_VSA1         -0.185686878 -0.1401755811 -0.1797429990 -0.162554832
#> EState_VSA10        -0.176873126 -0.1736311158 -0.1852959247 -0.191754991
#> EState_VSA11        -0.135785318 -0.1533179543 -0.1322332004 -0.143440885
#> EState_VSA2         -0.078456833 -0.0815849677 -0.0808870385 -0.070597237
#> EState_VSA3          0.051610202  0.0387716914  0.0338090349  0.029675238
#> EState_VSA4          0.061541955  0.0658025736  0.0782594319  0.069973067
#> EState_VSA5          0.102482614  0.0679973205  0.1012429352  0.088993776
#> EState_VSA6         -0.038370613 -0.0416882578 -0.0504905378 -0.052082238
#> EState_VSA7          0.154262704  0.1643795828  0.1625617768  0.169747667
#> EState_VSA8          0.013893786  0.0056598326  0.0083702455  0.016880202
#> EState_VSA9          0.166310211  0.1856616423  0.1606148394  0.190642377
#> ExactMolWt           0.053054280  0.0515050051  0.0483101136  0.053406840
#> FractionCSP3         0.000000000  0.0000000000  0.0000000000  0.000000000
#> HallKierAlpha        0.030514783  0.0000000000  0.0297625810  0.000000000
#> HeavyAtomCount       0.039341491  0.0385013808  0.0375811494  0.038632871
#> HeavyAtomMolWt       0.051733424  0.0490789338  0.0459055585  0.051109753
#> Ipc                  0.050632726  0.0637307437  0.0565392286  0.068551065
#> Kappa1               0.035304788  0.0383535696  0.0338690591  0.038113297
#> Kappa2               0.067712393  0.0745084697  0.0742019230  0.072672503
#> Kappa3               0.062257148  0.0701035050  0.0716493560  0.067746314
#> LabuteASA            0.050142624  0.0499968402  0.0478244167  0.050097794
#> MaxAbsEStateIndex   -0.039146432 -0.0543054885 -0.0406932648 -0.062893391
#> MaxEStateIndex      -0.039146432 -0.0543054885 -0.0406932648 -0.062893391
#> MinAbsEStateIndex    0.043067950  0.0563009812  0.0239362458  0.040734213
#> MinEStateIndex       0.071841372  0.0979753013  0.0851549095  0.090342357
#> MolMR                0.061725174  0.0625559599  0.0624998277  0.063868389
#> MolWt                0.053058569  0.0514544230  0.0482205043  0.053360274
#> NumValenceElectrons  0.039993620  0.0405481115  0.0387344903  0.040903509
#> PEOE_VSA1           -0.130282421 -0.1307716888 -0.1458147072 -0.123567630
#> PEOE_VSA10           0.049134011  0.0520587216  0.0508018665  0.063296861
#> PEOE_VSA11           0.040812175  0.0380754811  0.0551484133  0.035245452
#> PEOE_VSA12          -0.056157884 -0.0440904810 -0.0657240751 -0.057094977
#> PEOE_VSA13           0.065749616  0.0680695303  0.0591027397  0.068669623
#> PEOE_VSA14          -0.061012882 -0.0560020932 -0.0608975807 -0.057835132
#> PEOE_VSA2           -0.154870817 -0.1422264622 -0.1306981457 -0.148641790
#> PEOE_VSA3           -0.002719898  0.0000000000  0.0000000000  0.000000000
#> PEOE_VSA4            0.089147407  0.0737163606  0.0719140518  0.095487872
#> PEOE_VSA5            0.000000000  0.0000000000  0.0000000000  0.000000000
#> PEOE_VSA6            0.277611245  0.2839362031  0.2931765056  0.270559599
#> PEOE_VSA7            0.108821890  0.1320653897  0.1182601875  0.130198634
#> PEOE_VSA8           -0.247243320 -0.2696103723 -0.2696877592 -0.261724314
#> PEOE_VSA9            0.084293184  0.0728236326  0.0781935124  0.065455319
#> SMR_VSA1            -0.044119836 -0.0449981496 -0.0542551796 -0.044657484
#> SMR_VSA10            0.030824418  0.0308632096  0.0227546627  0.035556968
#> SMR_VSA2             0.000000000  0.0000000000  0.0000000000  0.000000000
#> SMR_VSA3            -0.205358204 -0.2159544798 -0.2069011100 -0.208245160
#> SMR_VSA4             0.078639516  0.1038665163  0.0981250398  0.117009422
#> SMR_VSA5             0.162588915  0.1699402964  0.1607963003  0.171922066
#> SMR_VSA6            -0.093637003 -0.0930114049 -0.0819349037 -0.107090140
#> SMR_VSA7             0.009420270 -0.0063707290  0.0006648859 -0.012901994
#> SMR_VSA9             0.163761244  0.1529862561  0.1763906625  0.151766396
#> SlogP_VSA1          -0.049917168 -0.0605197707 -0.0552257642 -0.048250040
#> SlogP_VSA10          0.111944852  0.0900613420  0.1008476608  0.091331590
#> SlogP_VSA11          0.106360448  0.0937367276  0.1187389215  0.098325427
#> SlogP_VSA12          0.113505170  0.1088267609  0.1021328634  0.114398253
#> SlogP_VSA2          -0.216375930 -0.2047689713 -0.2165113650 -0.213518073
#> SlogP_VSA3          -0.143330924 -0.1404977862 -0.1461786847 -0.147890841
#> SlogP_VSA4           0.061298715  0.0587429247  0.0853147669  0.077515646
#> SlogP_VSA5           0.188610259  0.2020480013  0.1944974403  0.204794425
#> SlogP_VSA6           0.132030292  0.1175436587  0.1189904140  0.117389988
#> SlogP_VSA7           0.074624909  0.0000000000  0.0573794453  0.000000000
#> SlogP_VSA8           0.000000000  0.0467841137  0.0503726078  0.069038943
#> TPSA                -0.236064621 -0.2249435140 -0.2315129410 -0.221493464
#> VSA_EState10         0.090816784  0.0908523155  0.0774317400  0.096850244
#> VSA_EState8          0.096054877  0.0685872814  0.0728417477  0.063003302
#> VSA_EState9         -0.103721352 -0.0873450239 -0.0920362740 -0.084475108
#>                             [,5]
#> BalabanJ            -0.063795033
#> BertzCT             -0.017152431
#> Chi0                 0.031811148
#> Chi0n                0.055495632
#> Chi0v                0.057136757
#> Chi1                 0.049730599
#> Chi1n                0.073753200
#> Chi1v                0.047452066
#> Chi2n                0.052360003
#> Chi2v               -0.006157849
#> Chi3n                0.042010090
#> Chi3v               -0.029485857
#> Chi4n                0.034413279
#> Chi4v               -0.035666154
#> EState_VSA1         -0.165372886
#> EState_VSA10        -0.189815878
#> EState_VSA11        -0.105312696
#> EState_VSA2         -0.104396636
#> EState_VSA3          0.057137069
#> EState_VSA4          0.075955048
#> EState_VSA5          0.095277540
#> EState_VSA6         -0.041141618
#> EState_VSA7          0.183088459
#> EState_VSA8          0.004411731
#> EState_VSA9          0.167072668
#> ExactMolWt           0.053478449
#> FractionCSP3         0.000000000
#> HallKierAlpha        0.037154710
#> HeavyAtomCount       0.039835234
#> HeavyAtomMolWt       0.051465121
#> Ipc                  0.066917004
#> Kappa1               0.037157515
#> Kappa2               0.071419409
#> Kappa3               0.061283699
#> LabuteASA            0.051240888
#> MaxAbsEStateIndex   -0.040498622
#> MaxEStateIndex      -0.040498622
#> MinAbsEStateIndex    0.047120502
#> MinEStateIndex       0.076039823
#> MolMR                0.064381592
#> MolWt                0.053412940
#> NumValenceElectrons  0.040813278
#> PEOE_VSA1           -0.133110675
#> PEOE_VSA10           0.080718892
#> PEOE_VSA11           0.033582225
#> PEOE_VSA12          -0.067229715
#> PEOE_VSA13           0.054946773
#> PEOE_VSA14          -0.072168100
#> PEOE_VSA2           -0.157359664
#> PEOE_VSA3           -0.039062321
#> PEOE_VSA4            0.100737619
#> PEOE_VSA5            0.000000000
#> PEOE_VSA6            0.276203551
#> PEOE_VSA7            0.151427628
#> PEOE_VSA8           -0.257153153
#> PEOE_VSA9            0.062040980
#> SMR_VSA1            -0.051088829
#> SMR_VSA10            0.025976829
#> SMR_VSA2             0.000000000
#> SMR_VSA3            -0.204315588
#> SMR_VSA4             0.102552683
#> SMR_VSA5             0.164540911
#> SMR_VSA6            -0.094161462
#> SMR_VSA7             0.003062854
#> SMR_VSA9             0.161720038
#> SlogP_VSA1          -0.059620056
#> SlogP_VSA10          0.110905386
#> SlogP_VSA11          0.104787988
#> SlogP_VSA12          0.103789870
#> SlogP_VSA2          -0.204715016
#> SlogP_VSA3          -0.159771140
#> SlogP_VSA4           0.078227964
#> SlogP_VSA5           0.203127851
#> SlogP_VSA6           0.124765010
#> SlogP_VSA7           0.000000000
#> SlogP_VSA8           0.062355929
#> TPSA                -0.237492929
#> VSA_EState10         0.075650779
#> VSA_EState8          0.098157228
#> VSA_EState9         -0.103764818

predict(fit, newx = x)

#>    [1] -0.296485961  0.058958465 -0.218142574  0.647341503 -0.653437464
#>    [6] -0.708343451  0.585936269 -0.502667747  0.033606089 -0.018455956
#>   [11]  0.819765576  0.134612180  0.983557012 -0.410453255  0.644106259
#>   [16]  0.874360765  0.046882173 -0.762943090 -0.128093521  0.522777807
#>   [21]  1.052398811  0.068794905  1.586978355  0.951615433  0.637293895
#>   [26]  0.604531153  1.208895716  0.718844819 -0.059338585 -0.311002666
#>   [31] -0.266750476  0.834126114  0.080441378 -0.606416016  0.682140745
#>   [36]  0.396389468  0.920289736 -0.341981971 -0.329755480  0.740624476
#>   [41] -0.357073263  0.953047861  0.625830585 -0.185082942 -0.001571340
#>   [46]  0.134236522  0.361759644  0.237763742 -0.207239439 -0.171671265
#>   [51] -0.004829744  0.049885685  1.375942905  1.067697052  0.035712371
#>   [56]  0.249387261  0.058040078  1.223857376 -0.092179083 -0.178428933
#>   [61]  1.564225945 -0.524849704  0.956017071  1.667356639  1.853968420
#>   [66]  1.179191803  0.301387071  1.308114414  0.868415233  0.534498683
#>   [71]  0.061518283 -0.072392829  0.667525095  0.531914444  0.609612100
#>   [76]  0.810768730  0.831428659  2.084248185  0.258113984  0.234777338
#>   [81]  0.633865659  0.245721983  0.396380963 -0.129126801 -0.010764383
#>   [86]  1.422485605  0.381267284  1.957330437  1.492349398  0.340926052
#>   [91] -2.944772831 -0.144298310  1.918097969  1.178075636  0.945099649
#>   [96] -0.004036703  0.458919625  3.037702955  1.453957630  0.729481340
#>  [101]  1.990105934  2.320598310  2.337954954  0.562875387  0.659474775
#>  [106]  0.343114767  2.624096931  0.752616827  1.876390686  2.103226989
#>  [111]  1.894969600  0.979671894  2.176538945  1.499658870 -0.865777815
#>  [116]  1.367280735  0.828101470  0.693970166  2.078740079  1.933680601
#>  [121]  2.211522718  0.809490865  1.459629306  2.093473206  0.754263661
#>  [126]  0.707251935  0.637107884  0.623570360  0.726237703  0.697071368
#>  [131]  1.654315721  0.448872286  0.753961422  0.818023291  3.180670079
#>  [136]  1.616416103  2.027083242  0.937626905  1.806414512  1.498345794
#>  [141]  1.959200103  1.566464984  2.402859087  2.213608801  1.621261915
#>  [146]  1.721835592  0.180082424  1.658558348  2.565570444  1.514422398
#>  [151]  2.316844818  1.182336782 -0.203481958  0.782367702  0.631297353
#>  [156]  2.869320008  1.531485387  1.179452759 -0.099188280  1.906778873
#>  [161]  1.702640630  0.195530966  0.812937941 -0.012808837  2.892234786
#>  [166]  2.326369689  3.389905741  1.692158946  0.076624409  3.179364263
#>  [171]  0.751578436 -0.203481958  2.220105717  1.383744415  0.913622336
#>  [176]  0.300216470  0.629607095  3.199125538  0.942164744  0.781026308
#>  [181]  1.097285107  1.182713669  2.549921892  0.291027165  2.063830786
#>  [186]  0.846879087  3.162799362  1.859844506  2.187633010  0.770810412
#>  [191]  2.434370294  2.787250012  1.075739242  2.715464227  2.523502125
#>  [196]  3.421449299  2.152031204  2.108038747  2.329580241  1.966978696
#>  [201]  2.810192074  2.001241430  1.354154131  3.176104942  2.577045816
#>  [206]  0.661852428  2.779711469  1.115176624  0.073908534  2.037131710
#>  [211]  0.063606757  2.478129562  2.478129562  1.093221472  0.234686245
#>  [216]  1.213798793  0.929922602  1.512833633  2.054399389  1.305560964
#>  [221]  1.685119244  1.285180155  1.483362222  2.881647028  2.622141632
#>  [226]  2.873779784  1.550728469  3.211598766  1.580334279  0.721959990
#>  [231]  2.910069248  0.923357627  0.623315442  2.023880814  2.900226162
#>  [236]  3.101497062  0.642286942  1.977090043  2.184533633  1.199902607
#>  [241] -0.160024919  1.060429697  2.540293791  1.230626477  1.037110558
#>  [246]  1.709751784  2.799906367  2.910815786  2.014800700  2.724996594
#>  [251]  0.389592987  1.982761497  1.826027381 -0.522197450  1.286284524
#>  [256]  2.702173858  0.502587424  3.067032462  0.236761959  1.453704578
#>  [261]  0.117537507  2.234736417  0.777744589  1.638246049  1.828213184
#>  [266]  1.150149069  3.385973460  0.791941397  3.606280628  1.930561363
#>  [271]  0.833617038  1.959796883  1.822075973  1.785616152  1.309471476
#>  [276]  3.567826949  1.211836305  0.894574059  2.187236322  0.462616881
#>  [281]  1.877466644 -0.322654763  1.169324080  0.486664473  2.502644691
#>  [286]  1.285775479  2.896337673  0.557700996  1.619979995  2.209455363
#>  [291]  2.128970760  1.102482943  3.177044999  1.658526464  1.295907146
#>  [296]  1.112635697  2.100705243  1.202183767 -2.934435514  3.246507882
#>  [301]  2.635074080  0.474185113  3.119095128  3.192988852  3.564723790
#>  [306]  2.986226295  1.258581989  1.070329629  3.392876079  1.797873836
#>  [311]  3.825818039  2.751786954  2.673851253  0.824300450  4.030342616
#>  [316]  2.017819475  1.653213601  3.014463264  2.797144331  3.217943517
#>  [321]  1.627932185  1.918965375  2.747186424  1.439878478  1.478688850
#>  [326]  1.906555767  2.333206180  1.422415210  1.361287313  1.017154488
#>  [331]  1.390147212  2.176619149  1.750439261  2.375737196  1.943951239
#>  [336]  3.649940659  2.251121021  1.571831278  1.225957752  1.673778346
#>  [341]  3.230594651  2.903851927  3.216225823 -0.311833118  3.417801536
#>  [346]  3.273769380  2.005526590  2.589906437  2.442985083  1.960058913
#>  [351]  1.878180842  0.414883564  2.218186373  1.820020226  1.994645875
#>  [356]  4.478588400  1.599500185  1.363774665  1.781609675  1.664320749
#>  [361]  1.865466101  2.142864687  3.485704341  1.999367845  1.845419921
#>  [366]  2.218186373  0.193869500  3.521485658  1.517736784  4.439131082
#>  [371]  2.090832852  1.991241822  3.546401000  1.969072463  1.753464849
#>  [376]  1.953746301  1.603125983  2.591467191  0.048500357  1.792925640
#>  [381]  1.311436039  2.391844541  1.444986939  1.596163562  2.163208244
#>  [386]  3.321827881 -0.421643796  3.312975403  1.828467327  2.073802065
#>  [391]  4.011020367  1.585581358  2.130605677  4.008782501  3.650981813
#>  [396]  3.360762297  1.908436754  2.217121750  1.807772062  1.475373030
#>  [401]  2.040800309  3.086712011  2.984212023  2.060545468  2.301990670
#>  [406]  0.903760950  1.695717392  2.618325766  3.116945948  3.520882681
#>  [411]  2.829257893  2.191928017  3.708371344  1.726448854  1.941989377
#>  [416]  1.787517563  3.534560079  1.280858538  1.979296622  3.626508564
#>  [421]  2.286499214  0.193950257  1.863219709  1.931837811  0.117126773
#>  [426]  1.632872949  4.686134329  5.087259829  5.637206788  2.236603488
#>  [431]  2.327677294  2.098220706  4.731952983  1.683544812  3.362411377
#>  [436]  1.908020227  2.179887709  1.629581709  1.841659401  1.629581709
#>  [441]  2.382614959  2.226273118  3.720399478  3.388839888  1.747677592
#>  [446]  2.817006227  2.549605180  2.585600528  2.352261137  2.586280011
#>  [451]  4.323275602  2.331889351  2.528380807  2.534195918  3.788026038
#>  [456]  2.585600528  2.575031122  2.630143808  3.292536908  1.959770743
#>  [461]  2.570562199  2.966238163  3.472670535  2.697402553  2.275508272
#>  [466]  1.971077610  4.811142072  2.957005467  3.195510722  2.361987023
#>  [471]  3.292543482  2.797574040  4.669860067  3.371251935  1.867813671
#>  [476]  3.007999074  2.785752030 -0.880493199  3.057536690  2.130228139
#>  [481]  4.289296680  3.565534097  4.676700653  4.248859241  2.231040133
#>  [486]  2.329812216  2.023753829  1.131641145  2.480992793  4.605942140
#>  [491]  3.444088064  3.330554441  2.359858640  2.142921394  2.084855443
#>  [496]  2.451018907  0.208100435  4.202454067  4.924197227  4.176137261
#>  [501]  4.506152987  3.206156767  3.295391987  3.344848000  2.845873319
#>  [506]  5.176216379  1.274241707  5.507663840  3.813903950  4.330431105
#>  [511] -0.215151035  3.142157482  3.773420650  3.216684808  2.978269904
#>  [516]  4.171804473  3.098855165  4.100281345 -0.354603902  3.087134446
#>  [521]  5.403956851  3.750234472  4.228947626  4.056762617  4.730406039
#>  [526]  0.974717906  4.296270139  3.591937766  6.353780512 -0.723653644
#>  [531]  6.770683032  4.015949132  5.197379451  3.609789270  6.939844122
#>  [536] -0.012658875 -1.090989100  0.104127211  0.365815908  0.472298961
#>  [541] -0.092818219  2.922161600  4.697409348  4.355078567  3.905960286
#>  [546]  4.275033791  4.158395653  5.381753415  0.566653460  0.241304015
#>  [551]  0.497336425  1.273012889 -0.332694298  3.572253455  5.205706751
#>  [556]  4.253506335  1.020515516 -0.832926346  0.973466731  2.834912956
#>  [561]  0.710371570  0.327219341  2.562252981  1.495570874  0.280209936
#>  [566]  3.447746336  2.688371609  0.750028714  3.714025366  2.524660375
#>  [571]  1.941851843  2.995770508  1.909906505  0.803512057  3.241789415
#>  [576]  1.784791198  3.170137126  2.278764986 -0.363188834  2.610943113
#>  [581]  2.422247857  1.101696129  0.116367973 -0.005988559  1.414345708
#>  [586] -0.277108917 -0.121910245  1.330495850  0.795500428  2.695065136
#>  [591]  2.259232030  1.622851126  0.011927835  3.906250028  1.481845296
#>  [596]  1.051899163  4.372205074  0.531315874  0.961502358  4.661444164
#>  [601]  1.516142187  1.410224204  3.151217814  1.431874560  1.454326927
#>  [606]  1.748005119 -0.012564085  1.490033801  1.472053860  2.845276385
#>  [611]  1.857123528  1.504486292  1.678419494  0.538477205  3.249680831
#>  [616]  3.201635505  1.887862456  1.507457587  3.364113054  1.386969271
#>  [621]  1.459860885  2.061808839  1.866500027  2.333673782  3.053234308
#>  [626]  1.339586117  0.167054500  1.102013661  1.400013633  1.073794661
#>  [631]  1.917288585  1.861466064  2.208115250  2.875151891  1.807564757
#>  [636]  2.629356701  2.574361710  3.889348297  3.246722357  2.616186305
#>  [641]  2.830418429  1.948750300  1.803987739  2.531205423  2.824977926
#>  [646]  3.492684608  1.605094081  3.973493843  2.225339983  2.107074944
#>  [651]  1.075983524  2.202425147  1.956465562  2.065991122  2.198966341
#>  [656]  2.441206339  1.131980917  2.537956708  3.232389357  2.038688693
#>  [661]  2.651477787  1.032965927  2.616373209  3.412980808  3.864520264
#>  [666]  3.547084476  2.619765036  2.383750565  2.562867689  2.283239225
#>  [671]  1.748502977  3.023659080  1.384357803  2.125025194  2.023609019
#>  [676]  2.125657758  1.868093166  2.499360747  1.816374558  1.628588748
#>  [681]  2.606194815  2.065574079  1.473296763  1.750980746  1.389730057
#>  [686]  1.180028965  2.249315613  2.397838763  2.539784733  2.807004329
#>  [691]  1.668426865  2.370018128  2.100265151  2.396625253  2.944013087
#>  [696]  2.238707999  3.427064163  2.019950832  1.896949525  2.146012296
#>  [701]  2.487951503  2.633829001  2.061919149  2.226657696  2.897286072
#>  [706]  3.618239280  3.613609686  3.633462810  3.525014073  2.941252029
#>  [711]  3.482269330  5.074295170  3.049222527  2.288560781  4.285285922
#>  [716]  4.954820734  5.161250947  0.914722789  2.604072893  2.375931965
#>  [721]  2.171830476  1.473784357  1.379145848  1.383084936  1.823604576
#>  [726]  1.361681760  1.470450853  0.764629669  3.377207705  4.320367585
#>  [731]  4.037680551  4.290424206  3.928462556  4.049209462  3.770403643
#>  [736]  3.419906281  3.486239104  3.597873014  4.321207604  4.298687721
#>  [741] -0.212927248  4.177590636  3.564443844  2.965260362  3.864567395
#>  [746]  3.554005245  0.550929499  2.935584658  2.885250150  3.745819311
#>  [751]  3.959480945 -0.459922130  0.981967012  3.810461507  1.464280254
#>  [756]  2.478852425  3.026423480  2.253615624  2.375111699  1.757738598
#>  [761]  2.003465357  1.786658581  0.451575994  1.150650755  2.638346769
#>  [766]  3.513111647  2.684340285  2.059725343  2.196049888  1.767474264
#>  [771]  3.262179474  2.257155446  2.140751335  2.936667264  2.836887589
#>  [776]  3.569439532  2.446708103  3.419381079  2.676785252  2.535281271
#>  [781]  1.241518228  3.404057933  2.631758092  1.108156518  1.539846577
#>  [786]  3.182895482  3.309548460  2.154333658  1.051368191  1.089287541
#>  [791]  2.442589265  1.141590543  2.483431311  2.166572694  1.139100041
#>  [796]  1.870081913  2.655986440  3.103348593  3.000524860  1.282704802
#>  [801]  2.760570046  2.074501147  1.866465964  2.248457559  2.731103128
#>  [806]  0.221744907  3.473647430  2.791790293  2.866156045  2.892760301
#>  [811]  3.368122236  3.329312387  3.614020090  2.639228742  3.709748650
#>  [816]  2.313643097  2.341037994  1.912000804  3.555079124  4.150469162
#>  [821]  3.500605757  4.533455721  0.254330686  3.679478773  3.810681661
#>  [826]  4.043124300  3.591562582  3.742275216  3.352745089  3.243322920
#>  [831]  4.294040983  1.777342192  3.261599894  2.866811020  3.930352254
#>  [836]  4.008544558  4.178434695  3.135922101  3.570477913  3.663841249
#>  [841]  3.541447979  5.092404951  3.387252741  1.642057250  2.953098241
#>  [846]  4.150300894  3.964409224  3.990066072  2.906325916  4.719046393
#>  [851]  3.928864423  4.236361390  4.508049608  4.038029202  4.075715995
#>  [856]  3.939517767  2.809910214  4.744015137  4.569715322  4.159101274
#>  [861]  4.365446813  4.825493468  1.845947416  4.296435538  4.781937392
#>  [866]  5.341954490  5.749077776  5.764739507  5.292965341  5.196691283
#>  [871]  0.076463967  0.083303126  0.070526674  0.368864064  0.399816968
#>  [876]  0.962743996  0.182728251 -1.025460283 -0.077009027  0.352711515
#>  [881] -0.568453953  1.666916051  2.153290354  2.485019516  2.748094657
#>  [886]  4.723304982  3.846799165  1.372664185  1.057173226  1.534220483
#>  [891]  1.636643992  1.725852381  0.373695475  2.325286846  1.949878176
#>  [896]  0.331257896  0.363917661  0.593056535  1.626659811  1.272001874
#>  [901]  1.657456751  1.679323198  1.898607158  1.555125526  2.284596047
#>  [906]  1.291728242  0.541486683  2.879338234  2.586945103  0.452254803
#>  [911]  0.424212272  0.865902032  0.604125001  0.669702368  0.883592601
#>  [916] -0.284617647 -0.195847922 -0.161384892  0.043388376  0.209648004
#>  [921]  0.584506594  0.905709499 -0.658752998 -0.463418448 -0.356183314
#>  [926] -0.319598366 -0.045172256  2.554708791  1.910745381  0.632452170
#>  [931]  0.796638795  3.906250028  0.487279426  1.558415293  4.756435593
#>  [936]  4.257893491  0.375091090 -0.363188834  1.501409587  1.941851843
#>  [941]  1.704211734  2.278764986  4.778858907  4.253506335  1.110565433
#>  [946]  1.280858538  0.143354271  2.422247857  0.794788670  2.199420904
#>  [951]  1.909906505  2.631199362  0.992616719  0.842084074  0.247062472
#>  [956]  1.098175197  0.015195862  2.010204489  0.462588591  1.622851126
#>  [961]  1.107207614  1.401263832  1.252014617  2.090160417  2.199420904
#>  [966]  0.982599251  2.192504246  3.592659709  3.066997040  2.985417919
#>  [971]  0.605149302  0.641367344  3.226854295  0.742923550 -0.027790971
#>  [976]  2.513659825  2.554708791  1.066768539  1.848334339  0.378640365
#>  [981]  0.508492838  0.398771136  1.156828692  4.645894274  0.797809445
#>  [986]  0.038650729  0.318615001  0.229019798  2.589871314 -0.178428933
#>  [991]  4.286473777  2.903851927  0.280209936  2.772139739  2.144110252
#>  [996]  1.124851023  0.556778183  1.302771294  0.153529945  0.636043413