Print enpls.ad Object

Print enpls.ad object.

# S3 method for enpls.ad
print(x, ...)

Arguments

x
An object of class enpls.ad.

...
Additional parameters for print.

See also

Examples

data("alkanes")
x = alkanes$x
y = alkanes$y

# training set
x.tr = x[1:100, ]
y.tr = y[1:100]

# two test sets
x.te = list(
  "test.1" = x[101:150, ],
  "test.2" = x[151:207, ])
y.te = list(
  "test.1" = y[101:150],
  "test.2" = y[151:207])

set.seed(42)
ad = enpls.ad(
  x.tr, y.tr, x.te, y.te,
  space = "variable", method = "mc",
  ratio = 0.9, reptimes = 50)
print(ad)

#> Model Applicability Domain Evaluation by ENPLS
#> ---
#> Absolute mean prediction error for each training set sample:
#>   [1]  0.97153796  0.16928726  0.05944468  1.15101213  0.18595071  1.05344651
#>   [7]  0.03070777  0.62199462  0.40126020  0.52913091  0.18560059  0.26322033
#>  [13]  3.54698709  0.68688498  0.17990776  3.24952244  0.85677422  0.08242658
#>  [19]  0.79632840  1.17930357  0.13160151  0.31043792  0.61310198  0.15200969
#>  [25]  0.46097347  0.70993067  0.01784780  0.70004480  0.73515927  0.45556212
#>  [31]  1.73849738  0.71673405  1.32057831  0.82665643  0.83401155  1.15883942
#>  [37]  0.49941872  1.17376284  1.31551629  1.10328094  2.23706412  2.31619155
#>  [43]  1.34794964  1.23047787  1.40788202  1.43837235  1.02639699  0.84919826
#>  [49]  2.05360492  2.09992008  1.53722054  0.01313521  3.80432838  1.01540426
#>  [55]  0.44732305  0.80394340  2.44567330  2.57179573  1.40607483  0.60492813
#>  [61]  0.83346464  3.44268153 14.24598415  5.92767290  0.75659936  1.08507877
#>  [67]  2.41609093  1.31320751  1.66881757  2.17913400  0.73531669  1.12455343
#>  [73]  1.09753983  2.21869835  1.52882960  2.08736327  0.36101730  1.67305229
#>  [79]  1.93241034  0.57684023  0.87210152  0.49228981  5.52717380  0.42720243
#>  [85]  2.60538928  0.84898793  2.12027511  1.91053568  1.49450208  0.38623837
#>  [91]  1.50140431  3.72459877  3.25992914  3.96635992  6.40761732  1.79428022
#>  [97]  3.54299617  1.91275597  1.68130657  0.97329019
#> ---
#> Prediction error SD for each training set sample:
#>   [1] 0.6110000 1.2055355 0.4920854 0.4668226 0.5889777 0.2820970 0.4757924
#>   [8] 0.2930795 0.4306813 0.2859090 0.3270101 0.2703124 0.7033181 0.3196315
#>  [15] 0.1452306 0.5084926 0.2946149 0.2212400 0.3262345 0.4022348 0.3160290
#>  [22] 0.2358490 0.4441658 0.3396711 0.2395216 0.4733301 0.3878655 0.2486117
#>  [29] 0.4630743 0.3894329 0.3119522 0.1871560 0.5039869 0.2272238 0.2076371
#>  [36] 0.2711488 0.4752279 0.3942441 0.3034399 0.2486551 0.5358166 0.2777923
#>  [43] 0.3232792 0.6274605 0.5146265 0.5449119 0.7525233 0.4102806 0.2982447
#>  [50] 0.2811808 0.3936532 0.2629753 0.3690240 0.3076202 0.2367177 0.4009293
#>  [57] 0.3577560 0.3120052 0.5393531 0.3256769 0.3923633 0.2127193 0.4353434
#>  [64] 0.2834852 0.3049661 0.5477951 0.3057302 0.3391954 0.3535890 0.4342044
#>  [71] 0.2293895 0.3025600 0.3356907 0.2621719 0.4133528 0.3990922 0.5048294
#>  [78] 0.5043581 0.3023483 0.4406910 0.4529266 0.1995952 0.3524072 0.3241378
#>  [85] 0.2641338 0.3638877 0.1743786 0.4585358 0.2305309 0.4455057 0.4442577
#>  [92] 0.3004639 0.8573116 0.5690498 0.2612492 0.2389716 0.3184314 0.4237129
#>  [99] 0.1989182 0.4820063
#> ---
#> Absolute mean prediction error for each test set sample:
#> [[1]]
#>  [1]  1.677055037  0.465530955  1.785101266  0.255388368  4.688402786
#>  [6]  0.423230157  0.777426854  1.916323815  3.385829593  2.747750195
#> [11]  2.255154570  2.872893638  3.647078436  2.030644531  0.016836164
#> [16]  2.409417166  3.734872401  0.001372733  2.517275677 15.728495321
#> [21] 16.502570071 14.376202512 17.914499911 11.082355854 12.643921072
#> [26] 41.279575600 16.421638519 15.349899295 15.110421273 10.527782122
#> [31]  6.230141083 15.520005323 15.410699050 16.324071428  7.023218103
#> [36] 16.806059960 15.204958375  7.517172460  5.440768333 16.320033561
#> [41]  9.644307141 14.995012611  7.760838660  5.358891705 16.642337155
#> [46] 10.876956089  6.980834787  5.733473821 10.984006203 11.043549585
#> 
#> [[2]]
#>  [1]  6.6468009  5.4191082 15.3310944 13.9633352  4.6661246  5.5203543
#>  [7]  4.3801550 11.1197332  4.3281840  2.7968025 47.3883886 40.5102593
#> [13] 44.1944866 41.5095080 44.4395707 39.3025487 49.4833591 43.5436261
#> [19] 32.4220973 63.4631188 37.6711394 31.9803220 35.1509916 43.9719265
#> [25] 39.5493049 31.2599918 28.5680772  0.1643899  1.1527738  3.4210308
#> [31]  2.6550668  2.0798862  1.6810499  0.6184867  6.3098807  1.4817469
#> [37]  5.2326925 22.7055916  0.7659122  1.4525070  3.8401601  0.9958045
#> [43]  4.9242235  0.8359972  4.2862191 20.1491873  6.9284571 17.2741198
#> [49]  6.0606433  9.5928046 15.3388224 14.1534032 14.7754516 15.6195331
#> [55] 10.4963665 16.4348771 14.4196109
#> 
#> ---
#> Prediction error SD for each test set sample:
#> [[1]]
#>  [1]   0.3580779   0.4773146   0.3395021   0.3919712   0.6965550   0.8134028
#>  [7]   0.3011729   0.5079211   0.5253101   0.6758042   0.5423055   0.6907619
#> [13]   0.5842912   0.6798572   0.6616702   0.8680762   0.7504019   0.8724157
#> [19]   1.0255729   9.6938114   9.4251146   9.1577657   9.0230936   7.6950427
#> [25]   6.5989738 108.7514002   8.3067098   8.2309732   7.9114093   8.3247198
#> [31]   8.3068948   7.5217637   7.7620859   8.0799532   8.2248186   7.9464715
#> [37]   7.6701586   8.0232303   8.0675082   7.3080276   7.9242841   7.8700847
#> [43]   8.0496976   8.0470546   7.7558405   7.8248017   7.8380557   7.8743126
#> [49]   7.8692260   7.6042991
#> 
#> [[2]]
#>  [1]   7.8891084   7.8460962   7.6458243   7.3618892   8.0642019   7.7154895
#>  [7]   7.6839861   6.6155478   7.6011514   7.6555093  24.0452207  20.7981116
#> [13]  20.9342276  22.3519817  21.5889352  21.7714139  22.1714377  22.1265447
#> [19]  21.9850561  94.8367059  21.8926680  21.8679147  21.1881839  21.6123203
#> [25]  21.6975435  21.4501399  21.3677250   0.2972805   0.6741695   0.8297756
#> [31]   0.5333837   0.6587137   0.3696752   0.4030205   0.5832984   0.1390277
#> [37]   0.3212491 114.8332814   0.2651159   0.3162890   0.2967644   0.1855057
#> [43]   0.3447145   0.3990924   0.7422411   9.2489665   8.1359606   8.6111165
#> [49]   8.2404925   8.2856108   8.3028043   6.9713116   8.0910335   8.0863983
#> [55]   8.1972618   8.2192021   8.0305364
#>