Print enpls.ad Object

Print enpls.ad object.

# S3 method for enpls.ad
print(x, ...)

Arguments

x
An object of class enpls.ad.

...
Additional parameters for print.

See also

Examples

data("alkanes")
x = alkanes$x
y = alkanes$y

# training set
x.tr = x[1:100, ]
y.tr = y[1:100]

# two test sets
x.te = list(
  "test.1" = x[101:150, ],
  "test.2" = x[151:207, ])
y.te = list(
  "test.1" = y[101:150],
  "test.2" = y[151:207])

set.seed(42)
ad = enpls.ad(
  x.tr, y.tr, x.te, y.te,
  space = "variable", method = "mc",
  ratio = 0.9, reptimes = 50)
print(ad)

#> Model Applicability Domain Evaluation by ENPLS
#> ---
#> Absolute mean prediction error for each training set sample:
#>   [1]  0.956860149  0.240525550  0.027768431  1.199237287  0.148366983
#>   [6]  1.060248987  0.013457809  0.607207868  0.352684070  0.500187536
#>  [11]  0.214690117  0.248146592  3.550828174  0.686001751  0.169838541
#>  [16]  3.245839634  0.872587748  0.080845084  0.780802118  1.207361654
#>  [21]  0.143988419  0.325331105  0.656828357  0.141244469  0.477280715
#>  [26]  0.760055246  0.006405968  0.714925803  0.784932531  0.449256239
#>  [31]  1.737793473  0.720461091  1.343551513  0.840673560  0.839911025
#>  [36]  1.162741134  0.461559847  1.182446411  1.294067916  1.123049439
#>  [41]  2.274835116  2.300408739  1.334823056  1.213408882  1.446446188
#>  [46]  1.400219452  1.077358779  0.844390485  2.070269753  2.103233811
#>  [51]  1.511601660  0.013295258  3.819553946  1.015609735  0.448721170
#>  [56]  0.811406653  2.443998340  2.552088332  1.386850832  0.622992790
#>  [61]  0.816456151  3.428279503 14.254061139  5.921244722  0.725880728
#>  [66]  1.134673576  2.412292095  1.276815736  1.665222319  2.164827431
#>  [71]  0.748959209  1.128759627  1.076630813  2.226143523  1.477811481
#>  [76]  2.118543699  0.324686148  1.684320651  1.931864200  0.572170166
#>  [81]  0.856709790  0.503634140  5.506608378  0.436364616  2.591539312
#>  [86]  0.842666023  2.122807103  1.904648492  1.489627959  0.382153792
#>  [91]  1.527585784  3.748404596  3.320959960  3.936815974  6.399167425
#>  [96]  1.793746562  3.537181304  1.938245894  1.677344936  0.944922196
#> ---
#> Prediction error SD for each training set sample:
#>   [1] 0.6317306 1.2576092 0.5522900 0.4618309 0.6372416 0.3029785 0.5165358
#>   [8] 0.2954675 0.4264100 0.2767761 0.3344779 0.2488248 0.7136464 0.3127953
#>  [15] 0.1670718 0.5144770 0.2964642 0.2382874 0.3543258 0.4008391 0.3324153
#>  [22] 0.2494126 0.4540302 0.3631917 0.2422918 0.5031328 0.3979549 0.2517021
#>  [29] 0.5177753 0.3855534 0.3182231 0.2076691 0.5212919 0.2330558 0.2073129
#>  [36] 0.2607434 0.5669314 0.3921931 0.3076910 0.2575816 0.5583562 0.2607333
#>  [43] 0.3049737 0.7423248 0.5207371 0.5225773 0.7312328 0.4353674 0.3021112
#>  [50] 0.2999804 0.3727870 0.2556193 0.3723387 0.3286442 0.2443668 0.4134501
#>  [57] 0.3747552 0.3025662 0.5027093 0.3120295 0.3591308 0.1971526 0.3972879
#>  [64] 0.2805786 0.2859100 0.5787721 0.2968229 0.3475461 0.3404194 0.4217507
#>  [71] 0.2215798 0.2925174 0.2998379 0.2705533 0.3943218 0.4138954 0.5349043
#>  [78] 0.5227050 0.2944736 0.4558113 0.4599414 0.1948032 0.3265541 0.3205197
#>  [85] 0.2608452 0.3891983 0.1626116 0.4790039 0.2470691 0.4705363 0.4069315
#>  [92] 0.3242741 0.8315532 0.5664095 0.2448131 0.2454349 0.3090163 0.4272330
#>  [99] 0.2011279 0.4878657
#> ---
#> Absolute mean prediction error for each test set sample:
#> [[1]]
#>  [1]  1.66708734  0.48966365  1.79196830  0.24778162  4.73069210  0.47024781
#>  [7]  0.76373582  1.87339818  3.39964547  2.79338479  2.22693487  2.92227835
#> [13]  3.66310337  2.05005468  0.06121133  2.33308748  3.79742430  0.05383003
#> [19]  2.57775425 15.82647980 16.54367731 14.45195072 17.94376572 11.14179875
#> [25] 12.66036398 38.66277069 16.44769003 15.36471409 15.09042729 10.53366039
#> [31]  6.18098990 15.68289923 15.41919948 16.35277103  6.99500560 16.83055735
#> [37] 15.18937352  7.52097634  5.39381926 16.47349286  9.70992233 15.03219626
#> [43]  7.77275521  5.33210822 16.68798701 10.91092222  6.99866608  5.68706039
#> [49] 10.97993261 11.14004566
#> 
#> [[2]]
#>  [1]  6.6817658  5.4070619 15.4023332 13.9963362  4.7554787  5.4973893
#>  [7]  4.3975610 11.1990944  4.3385374  2.8532293 47.4898781 40.5932027
#> [13] 44.2962060 41.5893639 44.5410573 39.3960097 49.5584567 43.5766156
#> [19] 32.4607176 60.9920219 37.7184148 31.9984564 35.2648999 44.0637285
#> [25] 39.6047911 31.3137170 28.6074976  0.1511898  1.1475888  3.3745254
#> [31]  2.6244120  2.0471295  1.6987472  0.6141990  6.3189577  1.4759353
#> [37]  5.2268485 20.1255597  0.7608450  1.4459932  3.8359917  0.9944998
#> [43]  4.8996446  0.8691618  4.3194447 20.1951276  7.1512509 17.2460690
#> [49]  6.0687412  9.5870258 15.3048113 14.1499572 14.7863430 15.5890983
#> [55] 10.5518465 16.4932683 14.4079678
#> 
#> ---
#> Prediction error SD for each test set sample:
#> [[1]]
#>  [1]  0.3450212  0.5171462  0.3149015  0.3960226  0.6753849  0.8264230
#>  [7]  0.3017017  0.5360031  0.5040856  0.6647788  0.5290552  0.6862329
#> [13]  0.5843001  0.6632770  0.6666389  0.8833116  0.7302581  0.8180692
#> [19]  1.0097590  9.1805285  8.9484260  8.6208357  8.5785840  7.2259387
#> [25]  6.3105603 99.8366893  7.7671557  7.7780971  7.4747776  7.8196536
#> [31]  7.8337923  6.8808153  7.2826441  7.5274870  7.7068523  7.5083545
#> [37]  7.2186851  7.5041859  7.5856556  6.6931212  7.3360340  7.3045169
#> [43]  7.4842652  7.5113914  7.3114598  7.2710148  7.3015978  7.3847274
#> [49]  7.3651111  7.0576724
#> 
#> [[2]]
#>  [1]   7.3169608   7.3134505   7.2020986   6.8803957   7.4883724   7.2183315
#>  [7]   7.1677281   6.2015494   7.0837130   7.0959864  22.6134212  19.5642227
#> [13]  19.6101749  20.9528754  20.1421309  20.3617094  20.6395574  20.6472320
#> [19]  20.5411473  87.9666580  20.4072192  20.4092229  19.8013036  20.0974708
#> [25]  20.1954576  20.0082232  19.9141551   0.3160014   0.6469758   0.8922060
#> [31]   0.5564348   0.6149711   0.3609206   0.3987404   0.5629879   0.1404934
#> [37]   0.3340906 104.9222603   0.2554147   0.3459965   0.3156373   0.1777127
#> [43]   0.3457124   0.3894733   0.7272909   8.7848807   7.4257367   8.1323931
#> [49]   7.7106461   7.7701878   7.8179235   6.6355067   7.5716199   7.5880908
#> [55]   7.5598082   7.5937495   7.4900812
#>