Print enpls.ad Object

Print enpls.ad object.

# S3 method for enpls.ad
print(x, ...)

Arguments

x
An object of class enpls.ad.

...
Additional parameters for print.

See also

Examples

data("alkanes")
x = alkanes$x
y = alkanes$y

# training set
x.tr = x[1:100, ]
y.tr = y[1:100]

# two test sets
x.te = list(
  "test.1" = x[101:150, ],
  "test.2" = x[151:207, ])
y.te = list(
  "test.1" = y[101:150],
  "test.2" = y[151:207])

set.seed(42)
ad = enpls.ad(
  x.tr, y.tr, x.te, y.te,
  space = "variable", method = "mc",
  ratio = 0.9, reptimes = 50)
print(ad)

#> Model Applicability Domain Evaluation by ENPLS
#> ---
#> Absolute mean prediction error for each training set sample:
#>   [1]  0.933956741  0.174824636  0.024280569  1.170418390  0.153135040
#>   [6]  1.076265538  0.002029500  0.629965136  0.371806761  0.510763165
#>  [11]  0.216289579  0.240917860  3.493910352  0.690537725  0.170287638
#>  [16]  3.198136356  0.850965664  0.062513536  0.755995622  1.171682806
#>  [21]  0.141895938  0.321023315  0.623504243  0.151459352  0.464318601
#>  [26]  0.736167599  0.004542222  0.706324645  0.768282784  0.431135808
#>  [31]  1.754660893  0.718575773  1.394784412  0.841094027  0.839062680
#>  [36]  1.153837468  0.472396578  1.153258092  1.321065714  1.109554055
#>  [41]  2.312295129  2.295773495  1.330044255  1.204205092  1.401988648
#>  [46]  1.415077449  1.116359169  0.791210168  2.095529097  2.126855923
#>  [51]  1.494085750  0.051076260  3.773702570  0.992903798  0.434182362
#>  [56]  0.787253759  2.455621515  2.512709129  1.385266789  0.645650296
#>  [61]  0.778074449  3.407957331 14.302981978  5.906056736  0.732195839
#>  [66]  1.119671815  2.445386904  1.280116401  1.713638698  2.161974614
#>  [71]  0.742675703  1.167791685  1.045598766  2.253394905  1.476225247
#>  [76]  2.101221444  0.344177960  1.689881219  1.967091024  0.603766764
#>  [81]  0.861057511  0.500570721  5.487320055  0.479293521  2.584291991
#>  [86]  0.852263220  2.094277693  1.932637842  1.512922272  0.347917974
#>  [91]  1.544533207  3.747137763  3.342190128  3.907526529  6.376483040
#>  [96]  1.787677298  3.510304757  1.973607466  1.671188871  0.972430242
#> ---
#> Prediction error SD for each training set sample:
#>   [1] 0.6020729 1.1109317 0.5709207 0.4788374 0.5799492 0.3041271 0.4403202
#>   [8] 0.2652545 0.3996291 0.2317465 0.3250620 0.2598722 0.6640500 0.3166683
#>  [15] 0.1744827 0.5062090 0.2735922 0.2383369 0.3581228 0.3597947 0.3258230
#>  [22] 0.2289847 0.4063172 0.3540365 0.2002951 0.4510573 0.3903182 0.2040832
#>  [29] 0.4927274 0.4006020 0.3234694 0.2058263 0.5260375 0.2593022 0.1887034
#>  [36] 0.2382415 0.5652798 0.3914926 0.2923978 0.2402676 0.5505460 0.2662303
#>  [43] 0.3086442 0.7318001 0.4426458 0.5081038 0.7721576 0.4578035 0.3384900
#>  [50] 0.3324957 0.3973875 0.2561214 0.3876233 0.3109661 0.2667963 0.4301501
#>  [57] 0.3693648 0.3578658 0.5101343 0.3503337 0.3967363 0.2304282 0.4420977
#>  [64] 0.2921458 0.2732953 0.5603824 0.3218922 0.3277753 0.4326325 0.4274726
#>  [71] 0.2076231 0.3005963 0.3446352 0.3163251 0.3886199 0.3966705 0.5201554
#>  [78] 0.5208405 0.3259874 0.4479218 0.4593539 0.1878024 0.3497973 0.3403728
#>  [85] 0.2659548 0.4048602 0.1756538 0.4430610 0.2550817 0.4707580 0.4203280
#>  [92] 0.3187072 0.8201504 0.5890336 0.2743511 0.2552916 0.3453270 0.4555029
#>  [99] 0.1994326 0.4502742
#> ---
#> Absolute mean prediction error for each test set sample:
#> [[1]]
#>  [1]  1.64303655  0.52011507  1.80854804  0.27397761  4.74956789  0.44629475
#>  [7]  0.79169171  1.91379941  3.35535996  2.81224380  2.23429390  2.89076073
#> [13]  3.63862526  2.06531267  0.02057142  2.37172067  3.77518783  0.06653200
#> [19]  2.54091920 17.04505989 17.74111346 15.63028636 19.07322256 12.14444444
#> [25] 13.47934542 26.66510076 17.56803835 16.40768031 16.12908791 11.63782760
#> [31]  7.27102422 16.62814072 16.44266661 17.45445217  8.08445076 17.83988220
#> [37] 16.20108244  8.59473896  6.46024697 17.38330300 10.75670060 16.10928232
#> [43]  8.84556663  6.40069954 17.66521305 11.94471131  8.04489715  6.72829346
#> [49] 12.03159357 12.12656206
#> 
#> [[2]]
#>  [1]  7.7276977  6.4515322 16.3416316 14.9570771  5.8148937  6.5178686
#>  [7]  5.4141705 12.0412141  5.3343060  3.8448102 50.5925090 43.2792335
#> [13] 47.0495280 44.4870893 47.3907448 42.2172564 52.5109564 46.5100741
#> [19] 35.3594474 51.3289567 40.6284944 34.9017403 37.9975198 46.9363116
#> [25] 42.4828167 34.1432397 31.4379864  0.1598946  1.2285287  3.4387596
#> [31]  2.5998681  2.0466946  1.6515350  0.5737215  6.4005967  1.4795442
#> [37]  5.2505980  7.0705691  0.7380260  1.4650057  3.8265141  0.9642647
#> [43]  4.9180170  0.8425861  4.4169974 21.3537888  8.1734897 18.3690604
#> [49]  7.1580738 10.6882846 16.3983330 15.0391111 15.8656240 16.6585201
#> [55] 11.6737871 17.6182522 15.4769111
#> 
#> ---
#> Prediction error SD for each test set sample:
#> [[1]]
#>  [1]   0.3596039   0.5174330   0.3132285   0.4025266   0.6773989   0.7645284
#>  [7]   0.2699342   0.4986699   0.5196797   0.6583661   0.5064671   0.6387197
#> [13]   0.5558690   0.6644286   0.5514427   0.7468375   0.6590804   0.8022008
#> [19]   0.9228705  10.5752370  10.2752284   9.9589388   9.8375996   8.3561926
#> [25]   7.2133433 122.0345508   8.9746241   8.9429357   8.5709829   9.0489806
#> [31]   8.9606253   8.0759433   8.4103867   8.7198550   8.8555998   8.6352266
#> [37]   8.3004547   8.6926192   8.6903975   7.8540780   8.5612989   8.4854419
#> [43]   8.6851603   8.6477987   8.4224476   8.4400320   8.4896967   8.4710084
#> [49]   8.5112512   8.2280276
#> 
#> [[2]]
#>  [1]   8.5145616   8.4351298   8.2979095   7.9663877   8.7604161   8.3078283
#>  [7]   8.2832620   7.1765371   8.1994560   8.2560986  26.1133479  22.5989017
#> [13]  22.7097367  24.2358951  23.3639503  23.5884342  23.9680502  23.9028739
#> [19]  23.8078600 109.8994900  23.6529433  23.6103401  22.9554590  23.3605597
#> [25]  23.4216036  23.2214309  23.0696251   0.2952354   0.7379650   0.7801094
#> [31]   0.5737839   0.6185236   0.3859655   0.4399115   0.6153185   0.1476588
#> [37]   0.3462843 127.5572600   0.2746705   0.3577757   0.3401878   0.2021822
#> [43]   0.3247426   0.3653422   0.7725961  10.0879845   8.7176687   9.3278581
#> [49]   8.8962413   8.9781230   8.9703347   7.5937235   8.7881098   8.7250009
#> [55]   8.8113154   8.8570657   8.6448928
#>